% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMEmod_indivplot.R
\name{ADME_IndPlt_server}
\alias{ADME_IndPlt_server}
\title{Server function for the concentration-time profile individual plots}
\usage{
ADME_IndPlt_server(id, adme_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in ADME_IndPlt_ui()}

\item{adme_args}{A Shiny reactive list with the output of modsol() and all shiny
parameters in pars()}
}
\value{
Server outputs for the concentration-time profile individual plots
drop down which includes four elements
}
\description{
This function generates the outputs defined in the ADME_IndPlt_ui()
function. This connects the download button, plot text, plot, and plot
caption to the elements that fill their spaces.
}
\seealso{
\code{\link[=plottingfunc_all]{plottingfunc_all()}},\code{\link[=plottingfunc_individual]{plottingfunc_individual()}}, \code{\link[=plt_arrange]{plt_arrange()}},
\code{\link[=caption_text]{caption_text()}}, which are all called by this function. \code{\link[=ADME_server]{ADME_server()}},
which calls this function.
}
