% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMEmod_multiplot.R
\name{ADME_MultPlt_ui}
\alias{ADME_MultPlt_ui}
\title{User interface function for the concentration-time profile multi-plot output}
\usage{
ADME_MultPlt_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in ADME_MultPlt_server()}
}
\value{
User interface for the multi-plot drop down which has three elements
}
\description{
This function outputs the user interface for the concentration-time profile
multi-plot drop down in the results card under the 'Run Simulation' tab.
The interface has three outputs: a download plot button, a plot, and a plot
caption.
}
