% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMEmod_main.R
\name{ADME_server}
\alias{ADME_server}
\title{Main server function to set up the 'Results' card under the concentration-time
profile module's 'Run Simulation' tab}
\usage{
ADME_server(id, pars, runsim)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in ADME_ui()}

\item{pars}{A list of all user input parameters for the entire app}

\item{runsim}{Action button titled 'Run Simulation' pressed by the user}
}
\value{
The four server outputs that fill the main concentration-time profile
user interface function consisting of two plots, two download buttons, and
one table of concentration-time profile results
}
\description{
#' @description
This function calculates the concentration-time profile's solution and outputs
the contents of the UI objects for the concentration-time profiles' 'Results'
card under the 'Run Simulation' tab. The contents of each UI object (drop
down) are then filled by four separate server modules.
}
\seealso{
\code{\link[=validate_text_ADME]{validate_text_ADME()}}, \code{\link[=modsol]{modsol()}}, \code{\link[=ADME_MultPlt_server]{ADME_MultPlt_server()}},
\code{\link[=ADME_IndPlt_server]{ADME_IndPlt_server()}}, \code{\link[=ADME_TCData_server]{ADME_TCData_server()}}, and \code{\link[=ADME_TKTable_server]{ADME_TKTable_server()}},
which are all called by this function
}
