% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMEmod_CTSimData.R
\name{ADME_TCData_server}
\alias{ADME_TCData_server}
\title{Server function for the concentration-time profile data download}
\usage{
ADME_TCData_server(id, adme_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in ADME_TCData_ui()}

\item{adme_args}{A Shiny reactive list with the output of modsol() and all shiny
parameters in pars()}
}
\value{
Server outputs for the concentration-time profile data download drop
down which includes two elements
}
\description{
This function generates the outputs defined in the ADME_TCData_ui()
function. This connects the two download buttons with the data to download.
}
