% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMEmod_CTSimData.R
\name{ADME_TCData_ui}
\alias{ADME_TCData_ui}
\title{User interface function for the concentration-time profile data download}
\usage{
ADME_TCData_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in ADME_TCData_server()}
}
\value{
User interface for the data download drop down with two download buttons
}
\description{
#' #' @description
This function outputs the user interface for the concentration-time profile
data download drop down in the results card under the 'Run Simulation' tab.
The interface has two outputs: a download button for concentration-time profile
data and a download button for simulation parameters.
}
