% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMEmod_TKsummary.R
\name{ADME_TKTable_server}
\alias{ADME_TKTable_server}
\title{Server function for the concentration-time profile toxicokinetic summary
table output}
\usage{
ADME_TKTable_server(id, adme_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in ADME_TKTable_ui()}

\item{adme_args}{Shiny reactive list with the output of modsol() and all shiny
parameters in pars()}
}
\value{
Server outputs for the concentration-time profile toxicokinetic summary
drop down which includes three outputs
}
\description{
This function generates the outputs defined in the ADME_TKTable_ui()
function. This connects the download button, data table, and caption defined
in the UI to the outputs that fill those spots.
}
\seealso{
\code{\link[=ADME_server]{ADME_server()}}, which calls this function
}
