% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMEmod_TKsummary.R
\name{ADME_TKTable_ui}
\alias{ADME_TKTable_ui}
\title{User interface function for the concentration-time profile toxicokinetic
summary table output}
\usage{
ADME_TKTable_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in ADME_TKTable_server()}
}
\value{
User interface for the toxicokinetic summary table drop down with
three elements
}
\description{
This function outputs the user interface for the concentration-time profile
toxicokinetic table drop down in the results card under the 'Run Simulation' tab.
The interface has three outputs: a download button to download the table, the
table of toxicokinetic summary statistics, and a table caption.
}
\seealso{
\code{\link[=ADME_ui]{ADME_ui()}}, which calls this function
}
