% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMEmod_main.R
\name{ADME_ui}
\alias{ADME_ui}
\title{Main user interface function to set up the 'Results' card under the
concentration-time profiles module's 'Run Simulation' tab}
\usage{
ADME_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in ADME_server()}
}
\value{
Four user interface drop down tabs identifying space for two plots, two
download buttons, and one table of concentration-time profile results.
}
\description{
#' @description
This function outputs the structure of the user interface for the concentration-
time profile 'Results' card under the 'Run Simulation' tab. Four drop down
tabs appear: a single plot with all compound concentrations in all model
compartments, multiple plots with one plot per compound, a download option for
users to download time course data and simulation parameters, and a table with
toxicokinetic summary statistics (Cmax, Tmax, AUC). The user interface contents
of each drop down is then filled by a separate UI module.
}
