% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{AP_Bioavailability}
\alias{AP_Bioavailability}
\title{User interface for the 'Bioavailability' card on the 'Advanced Parameters' tab}
\usage{
AP_Bioavailability()
}
\value{
The 'Bioavailability' card which offers users the option to customize
five bioavailability parameters, unless the parameter calculations module
function is chosen
}
\description{
User interface for the 'Bioavailability' card on the 'Advanced Parameters' tab
}
