% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{AP_ModelConditions}
\alias{AP_ModelConditions}
\title{User interface for the 'Model Conditions' card on the 'Advanced Parameters'
tab}
\usage{
AP_ModelConditions(ic_names, ic_comps)
}
\value{
The 'Model Conditions' card which allows the user to select various
different model conditions depending upon the module function chosen
}
\description{
User interface for the 'Model Conditions' card on the 'Advanced Parameters'
tab
}
