% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{AP_ModelSolver}
\alias{AP_ModelSolver}
\title{User interface for the 'Model Solver' card on the 'Advanced Parameters' tab}
\usage{
AP_ModelSolver()
}
\value{
The 'Model Solver' card which allows the user options to select various
solver conditions if the concentration-time profiles module function is chosen
}
\description{
This function outputs the user interface for the 'Model Solver' card on the
'Advanced Parameters' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=selectInput_ODEmethod]{selectInput_ODEmethod()}},
\code{\link[=numericInput_SolSteps]{numericInput_SolSteps()}}, \code{\link[=sliderInput_RTol]{sliderInput_RTol()}}, and \code{\link[=sliderInput_ATol]{sliderInput_ATol()}},
which are called by the current function
}
