% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{AP_OutputSpecification}
\alias{AP_OutputSpecification}
\title{User interface for the 'Output Specification' card on the 'Advanced Parameters'
tab}
\usage{
AP_OutputSpecification()
}
\value{
The 'Output Specification' card which offers the users different
selection options regarding the output for different module functions chosen
}
\description{
User interface for the 'Output Specification' card on the 'Advanced Parameters'
tab
}
