% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Not to be called directly by the user - app server function}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Input objects from the user interface}

\item{output}{Output objects from the server}

\item{session}{Session identifier}
}
\value{
Various server outputs
}
\description{
This function is the top-level server function for the ToCS application and
dictates the outputs to be displayed in the app_ui function. Calls to all other
server functions stem from this function. The current function calls the following
internal functions: Model_Input(), PreloadComps_UI(), CompileCompLst(),
UpdateInputs(), GatherInputVars()], InputRules_Children(), InputRules_Parents(),
and Run_Simulation().
}
