% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_BERplot.R
\name{BER_Plot_server}
\alias{BER_Plot_server}
\title{Server function for the BER plot}
\usage{
BER_Plot_server(id, ivive_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in BER_Plot_ui()}

\item{ivive_args}{A Shiny reactive list with the output of IVIVEsol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
Server outputs for the BER plot drop down which includes three elements
}
\description{
This function generates the outputs defined in the IVIVE_Table_ui()
function. This connects the download buttons with the plot to download,
generates the plot content, and creates the text for the plot caption.
}
\seealso{
\code{\link[=BERplotting]{BERplotting()}}, which is called by this function, and \code{\link[=IVIVE_server]{IVIVE_server()}},
which calls the current function
}
