% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_BERplot.R
\name{BER_Plot_ui}
\alias{BER_Plot_ui}
\title{User interface function for the BER plot}
\usage{
BER_Plot_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in BER_Plot_server()}
}
\value{
User interface for the BER plot drop down with three elements
}
\description{
This function outputs the user interface for the BER plot drop down in the
results card under the 'Run Simulation' tab. The interface has three outputs:
a plot download button, a plot with BER values, and a plot caption.
}
\seealso{
\code{\link[=IVIVE_ui]{IVIVE_ui()}}, which calls the current function
}
