% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_BERTable.R
\name{BER_Table_server}
\alias{BER_Table_server}
\title{Server function for the IVIVE dose table}
\usage{
BER_Table_server(id, ivive_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in BER_Table_ui()}

\item{ivive_args}{A Shiny reactive list with the output of IVIVEsol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
Server outputs for the BER table drop down which includes three elements
}
\description{
#' @description
This function generates the outputs defined in the BER_Table_ui()
function. This connects the download buttons with the table to download,
fills the data table with content, and creates the text for the table caption.
}
\seealso{
\code{\link[=IVIVE_server]{IVIVE_server()}}, which calls the current function
}
