% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_BERTable.R
\name{BER_Table_ui}
\alias{BER_Table_ui}
\title{User interface function for the BER table}
\usage{
BER_Table_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in BER_Table_server()}
}
\value{
User interface for the BER table drop down with three elements
}
\description{
#' @description
This function outputs the user interface for the BER table drop down in the
results card under the 'Run Simulation' tab. The interface has three outputs:
a table download button, a table with calculated BERs, and a table caption.
}
