% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{BERplotting}
\alias{BERplotting}
\title{Plot the bioactivity exposure ratio (BER) for all simulated chemicals}
\usage{
BERplotting(BERdata)
}
\arguments{
\item{BERdata}{A data frame of chemical names and BER values}
}
\value{
A ggplot plotting object with all BERs plotted and a red line indicating
the chemical prioritization cutoff
}
\description{
This function plots the bioactivity exposure ratios (BERs) for a list of
chemicals. The compounds are arranged in ascending order and then plotted using
ggplot2. A red line is also plotted at BER = 1, indicating the threshold where
compounds below the red line should be prioritized for further assessment.
}
