% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{Bind_Chem_Data}
\alias{Bind_Chem_Data}
\title{Title}
\usage{
Bind_Chem_Data(pars, pars_df)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}

\item{pars_df}{}
}
\description{
Title
}
