% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{Calc_OEDBER_RS_True}
\alias{Calc_OEDBER_RS_True}
\title{Calculate the oral equivalent doses and bioactivity exposure ratio when OED
samples are returned}
\usage{
Calc_OEDBER_RS_True(n, pars, bioactive_conc, exposuredata)
}
\arguments{
\item{n}{The number of compounds being simulated}

\item{pars}{A list of all user input parameters for the entire app}

\item{bioactive_conc}{A data frame with chemical names, their CAS numbers, and
their bioactive concentrations}

\item{exposuredata}{A data frame with chemical names, their CAS numbers, and
an upper, median, and lower exposure estimate}
}
\value{
A list with two elements: an array with oral equivalent doses and
a data frame with bioactivity exposure ratios
}
\description{
This function calculates an array of oral equivalent dose (OED) samples and the 5th
quantile OED as well a data frame of bioactivity exposure ratios (BERs). The
BER is calculated using the 5th quantile OED. If no exposure data is uploaded
by the user, then no bioactivity exposure ratios are returned.
}
