% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_outputfuncs.R
\name{CalcAnalyticCss}
\alias{CalcAnalyticCss}
\title{Calculate the analytical steady state concentration of a chemical}
\usage{
CalcAnalyticCss(pars, i)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}

\item{i}{Chemical index on the simulated chemicals list}
}
\value{
A numerical value representing the analytical steady state concentration
in the specified units
}
\description{
This function calculates the analytical steady state concentration of an
individual compound using all user-selected parameters. The 'httk' function
calc_analytic_css is used to generate the output.
}
\seealso{
\code{\link[=SS_sol]{SS_sol()}}, which calls the current function
}
