% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_outputfuncs.R
\name{CalcClearance}
\alias{CalcClearance}
\title{This function calculates the total plasma clearance of a chemical}
\usage{
CalcClearance(pars, i)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}

\item{i}{A numerical value representing the index number of the compound to
simulate}
}
\value{
A numerical value, which is the plasma total clearance for compound i
}
\description{
#' @description
This function calculates the total plasma clearance of a compound using the
'httk' package's 'calc_total_clearance' function.
}
\seealso{
\code{\link[=Parsol]{Parsol()}}, which calls the current function
}
