% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_outputfuncs.R
\name{CalcCssDay}
\alias{CalcCssDay}
\title{Calculate the day steady state is reached}
\usage{
CalcCssDay(pars, i)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}

\item{i}{Chemical index on the simulated chemicals list}
}
\value{
A list composed of the average concentration at the end of the
simulation, the fraction of the true steady state reached on the steady state
day, the maximum concentration of the simulation, and the day steady state
was reached
}
\description{
This function calculates the days to steady state of an individual compound
as well as other related steady state attributes using all user-selected
parameters. The 'httk' function calc_css is used to generate the output.
}
\seealso{
\code{\link[=SS_sol]{SS_sol()}}, which calls the current function
}
