% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_outputfuncs.R
\name{CalcElimRate}
\alias{CalcElimRate}
\title{Calculate the elimination rate of a chemical}
\usage{
CalcElimRate(pars, i)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}

\item{i}{A numerical value representing the index number of the compound to
simulate}
}
\value{
A numerical value, which is the elimination rate for compound i
}
\description{
This function calculates the elimination rate of a compound using the 'httk'
package's 'calc_elimination_rate' function.
}
