% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_outputfuncs.R
\name{CalcHalfLife}
\alias{CalcHalfLife}
\title{Calculate the half-life of a chemical}
\usage{
CalcHalfLife(pars, i)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}

\item{i}{A numerical value representing the index number of the compound to
simulate}
}
\value{
A numerical value, which is the half-life for compound i
}
\description{
#' @description
This function calculates the half-life of a compound using the 'httk' package's
'calc_half_life' function.
}
