% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{CalcOED}
\alias{CalcOED}
\title{Calculate the oral equivalent dose (OED) of a chemical}
\usage{
CalcOED(i, pars, bioactive_df)
}
\arguments{
\item{i}{Index indicating which chemical on the list to simulate}

\item{pars}{A list of all user input parameters for the entire app}

\item{bioactive_df}{A data frame with the chemical name, CAS number, and
bioactive concentration (in uM units) for each chemical to simulate}
}
\value{
A numeric value or vector of values, depending on the parameter
}
\description{
This function calculates the oral equivalent dose (OED) of a chemical given
its bioactive concentration. The 'httk' package function 'calc_mc_oral_equiv"
is used to generate either a single OED or a vector of OEDs depending on
whether the user wants only a specific quantile returned or all OED samples
returned.
}
