% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_outputfuncs.R
\name{CalcPCs}
\alias{CalcPCs}
\title{Calculate the partition coefficients of a chemical}
\usage{
CalcPCs(pars, i)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}

\item{i}{A numerical value representing the index number of the compound to
simulate}
}
\value{
A list, which consists of the partition coefficients for compound i
}
\description{
This function calculates the tissue to unbound plasma partition coefficients
of a given chemical. This is done using the 'httk' package's
'predict_partitioning_schmitt' function.
}
\seealso{
\code{\link[=Parsol]{Parsol()}}, which calls the current function
}
