% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_outputfuncs.R
\name{CalcVdist}
\alias{CalcVdist}
\title{Calculate the volume of distribution of a chemical}
\usage{
CalcVdist(pars, i)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}

\item{i}{A numerical value representing the index number of the compound to
simulate}
}
\value{
A numerical value, which is the volume of distribution for compound i
}
\description{
#' @description
This function calculates the volume of distribution of a compound using the
'httk' package's 'calc_vdist' function.
}
