% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_ValidationText.R
\name{caption_text}
\alias{caption_text}
\title{Create the text for a figure/table caption}
\usage{
caption_text(func, model)
}
\arguments{
\item{func}{User selection with the input$func ID; either "Concentration-time
profiles", "Steady state concentrations", "In vitro in vivo extrapolation
(IVIVE)", and "Parameter calculations"}

\item{model}{User selection with the input$model ID; either "3compartmentss",
"Schmitt","1compartment","3compartment", "pbtk" or "fetal_pbtk"}
}
\value{
A string with text describing the returned concentration of the model.
}
\description{
Create the text for a figure/table caption
}
\seealso{
\code{\link[=ADME_MultPlt_server]{ADME_MultPlt_server()}} and \code{\link[=ADME_IndPlt_server]{ADME_IndPlt_server()}}, which calls the
current function.
}
