% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{checkboxInput_Log}
\alias{checkboxInput_Log}
\title{Generates a checkbox for the user to click if the user wants a log10 y-axis
on their respective plots}
\usage{
checkboxInput_Log(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A checkbox input
}
\description{
Generates a checkbox for the user to click if the user wants a log10 y-axis
on their respective plots
}
\examples{
checkboxInput_Log("logscale")

}
\seealso{
\code{\link[=RS_Actions]{RS_Actions()}}, which calls the current function
}
