% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{CompileCompLst}
\alias{CompileCompLst}
\title{Compile a list of all chemicals to simulate}
\usage{
CompileCompLst(
  func,
  spec,
  defaulthuman,
  model,
  insilico,
  preloadcomps,
  file,
  honda
)
}
\arguments{
\item{func}{User selection with the input$func ID; either "Concentration-time
profiles", "Steady state concentrations", "In vitro in vivo extrapolation
(IVIVE)", and "Parameter calculations"}

\item{spec}{User selection with the input$spec ID; either "Dog","Human","Mouse",
"Rabbit","Rat"}

\item{defaulthuman}{User selection with input$defaulttoHuman; either "Yes" for
using human in vitro data in place of animal data or "No" for only using
animal data}

\item{model}{User selection with the input$model ID; either "3compartmentss",
"Schmitt","1compartment","3compartment", "pbtk" or "fetal_pbtk"}

\item{insilico}{User selection with the input$insilicopars ID; either
"Yes, load in silico parameters" or "No, do not load in silico parameters"}

\item{honda}{User selection with the input$HondaIVIVE ID; either "NULL","Honda1",
"Honda2","Honda3","Honda4"}
}
\value{
A data frame with a single column of compound names in alphabetical
order
}
\description{
This function creates a data frame of chemical names in alphabetical order to
simulate. It checks the function's input parameters for validity. Then, if
applicable, the user's uploaded file of chemical data is uploaded and merged
any preloaded chemicals selected by the user.
}
