% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{CompilePars}
\alias{CompilePars}
\title{Parse together the input ID name of a user selection}
\usage{
CompilePars(VarName, input)
}
\arguments{
\item{VarName}{The name of the parameter}

\item{input}{The 'input' object used as the prefix for all ID's (should be
entered as 'input')}
}
\value{
A value for the given parameter
}
\description{
Parse together the input ID name of a user selection
}
