% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{ConvertBioactive}
\alias{ConvertBioactive}
\title{Convert a bioactive concentration from a nominal concentration to a free
concentration in vitro}
\usage{
ConvertBioactive(pars, bioactive_df)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}

\item{bioactive_df}{A data frame with the chemical name, CAS number, and
bioactive concentration (in uM units) for each chemical to simulate}
}
\value{
A data frame with the chemical name, CAS number, and bioactive
concentration
}
\description{
This function converts nominal bioactive concentrations to a free concentration
in vitro, if the user selects the 'Honda1' IVIVE assumption. If chosen, this
assumption uses the 'httk' package's 'armitage_eval' function to perform the
conversion. If the 'Honda1' assumption is not chosen, then the original
'bioactive_df' data frame is returned.
}
