% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{Create_Plotting_df}
\alias{Create_Plotting_df}
\title{Reformat the ADME solution array for a specific compartment}
\usage{
Create_Plotting_df(i, n, sol_array, chemnames, numtimes)
}
\arguments{
\item{i}{A number representing the index of the current model state to be plotted}

\item{n}{A number representing the number of compounds being plotted}

\item{sol_array}{An array (row = time steps, column = model state, sheet = compound) of the model solution from other code}

\item{chemnames}{A vector containing the names of the compounds being simulated}

\item{numtimes}{A vector containing the times the solution is evaluated}
}
\value{
A data frame with concentration-time profile data for compartment
i and it contains three columns: time, yvalues, and compound.
}
\description{
This function rearranges the concentration-time profile solution array (where
each page represents a compound, each column a model compartment, and each
row a time point) into a data frame with three columns (time, yvalues, and
compound name) for the ith model compartment.
}
\seealso{
\code{\link[=plottingfunc_all]{plottingfunc_all()}}, which calls this function
}
