% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{CS_PreloadedCompounds}
\alias{CS_PreloadedCompounds}
\title{User interface for the 'Preloaded Compounds' card on the 'Compound Selection'
tab}
\usage{
CS_PreloadedCompounds()
}
\value{
The 'Preloaded Compounds' card that allows the user to select desired
chemicals to simulate from a list as well as any chemical-related simulation
preferences
}
\description{
This function outputs the user interface for the 'Preloaded Compounds' card on the
'Compound Selection' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=selectInput_HondaCond]{selectInput_HondaCond()}}
and \code{\link[=selectInput_CompPreference]{selectInput_CompPreference()}}, which are called by the current function
}
