% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{CS_UploadedData}
\alias{CS_UploadedData}
\title{User interface for the 'Uploaded Data' card on the 'Compound Selection' tab}
\usage{
CS_UploadedData()
}
\value{
The 'Uploaded Data' card which gives the user the option to upload a
file with new chemicals and their data. If the IVIVE module is chosen, then
users can also upload bioactivity and exposure data.
}
\description{
User interface for the 'Uploaded Data' card on the 'Compound Selection' tab
}
