% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{DataInputType}
\alias{DataInputType}
\title{Are the entries in the uploaded compound data file the correct kind of data?}
\usage{
DataInputType(file, column_names)
}
\arguments{
\item{file}{The data frame with compound physical-chemical data uploaded by
the user}

\item{column_names}{A vector of the column names of the 'file' data frame}
}
\value{
If any data type in 'file' is not an appropriate data type for that
column, then this function returns a message detailing that there's an incorrect
data type. If all data types are appropriate in the 'file', then no message
is outputted.
}
\description{
This function checks if the data types that are in the uploaded compounds file
are appropriate for each column of data (i.e. character, numeric, logical).
}
\seealso{
\code{\link[=UploadComps_Check]{UploadComps_Check()}}, which calls the current function, and \code{\link[=DataTypeList]{DataTypeList()}}
and \code{\link[=type_func]{type_func()}}, which the current function calls
}
