% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{DataTypeList}
\alias{DataTypeList}
\title{Generate a list of acceptable input data types for the compound file user
upload}
\usage{
DataTypeList()
}
\value{
A list of all acceptable data types for each column in the 'httk'
package's chem.physical_and_invitro.data
}
\description{
This function produces a list of acceptable data types for each column of data
in the user's uploaded compound file.
}
\seealso{
\code{\link[=DataInputType]{DataInputType()}}, which calls the current function
}
