% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{Dosing_Output}
\alias{Dosing_Output}
\title{Turn dosing parameters into solve_model input format}
\usage{
Dosing_Output(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A list of dosing parameters (initial.dose, doses.per.day, daily.dose,
and dosing.matrix)
}
\description{
Turn dosing parameters into solve_model input format
}
