% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{DosingPar}
\alias{DosingPar}
\title{Return the list of dosing parameters needed for the ADME dosing parameter in
httk's solve_model function}
\usage{
DosingPar(
  dosenum,
  initdose,
  multdose,
  multdosetime,
  multdoseamount,
  multdoseodd
)
}
\arguments{
\item{dosenum}{The user's selection of the number of doses to administer,
either "Single dose" or "Multiple doses"}

\item{initdose}{A positive number; The user's input of the amount of initial
dose to administer}

\item{multdose}{The user's selection of the kind of multiple dosing to
simulate, either "Yes" or "No"}

\item{multdosetime}{A positive number; The user's slider selection of how
often to administer reoccurring doses}

\item{multdoseamount}{A positive number; The user's selection of the amount
of dose to give during evenly spaced intervals}

\item{multdoseodd}{A list of numbers; The user's text input of a list of times
to administer dose and the dose to administer at each time}
}
\value{
A list of dosing information taken in by httk's solve_model function.
The list has five entries: initial.dose, doses.per.day, daily.dose,
dosing.matrix, and forcings.
}
\description{
Return the list of dosing parameters needed for the ADME dosing parameter in
httk's solve_model function
}
