% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{ExposureUpload_Check}
\alias{ExposureUpload_Check}
\title{Does the uploaded exposure file have the correct formatting and data type
as well as no missing required data?}
\usage{
ExposureUpload_Check(value, input)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}
}
\value{
If the uploaded exposure file is missing needed data or has incorrect
formatting, then this function returns a message detailing the problem. If all
aspects of the exposure file are appropriate, then no message is outputted.
}
\description{
Does the uploaded exposure file have the correct formatting and data type
as well as no missing required data?
}
