% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{fileInput_BioactiveConc}
\alias{fileInput_BioactiveConc}
\title{Generates a browser and file upload box for the user to upload a CSV file of
bioactivity data for the selected compounds during the in vitro in vivo
extrapolation (IVIVE) simulation}
\usage{
fileInput_BioactiveConc(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A browser and file upload box
}
\description{
Generates a browser and file upload box for the user to upload a CSV file of
bioactivity data for the selected compounds during the in vitro in vivo
extrapolation (IVIVE) simulation
}
\examples{
fileInput_BioactiveConc("BioactiveFile")

}
\seealso{
\code{\link[=CS_UploadedData]{CS_UploadedData()}}, which calls the current function
}
