% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{fileInput_ExposureData}
\alias{fileInput_ExposureData}
\title{Generates a browse and upload box for users to upload a CSV file of exposure
data for in vitro in vivo extrapolation (IVIVE) simulations}
\usage{
fileInput_ExposureData(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A browse and file upload box
}
\description{
Generates a browse and upload box for users to upload a CSV file of exposure
data for in vitro in vivo extrapolation (IVIVE) simulations
}
\examples{
fileInput_ExposureData("fileExposure")

}
\seealso{
\code{\link[=CS_UploadedData]{CS_UploadedData()}}, which calls the current function
}
