% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{FillExposureData}
\alias{FillExposureData}
\title{Replace NAs in exposure data frame with values}
\usage{
FillExposureData(exposure_df)
}
\arguments{
\item{exposure_df}{A data frame with chemical names and exposure estimates. '}
}
\value{
A data frame with chemical names and exposure estimates.
There should be no NA values in the outputted data frame.
}
\description{
This function fills in the NA values in the exposure data frame by placing a
small deviation (1e-8) from the non-NA elements of the data frame for that
particular chemical. This is done so that the exposure data can be plotted as
line and point ranges.
}
\seealso{
\code{\link[=OEDSample_Exposure_Plot]{OEDSample_Exposure_Plot()}} and \code{\link[=OEDPoint_Exposure_Plot]{OEDPoint_Exposure_Plot()}}, which
calls the current function
}
