% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{GatherInputVars}
\alias{GatherInputVars}
\title{Compile all parameters into one final list}
\usage{
GatherInputVars(input, compoundlist)
}
\arguments{
\item{input}{The 'input' object used as the prefix for all ID's (should be
entered as 'input')}

\item{compoundlist}{A data frame with a single column that has the alphabetized
list of compounds to simulate}
}
\value{
A named list of all parameters in the GUI to be used in the main
module functions
}
\description{
Compile all parameters into one final list
}
