% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{getCASnums}
\alias{getCASnums}
\title{Generate a vector of CAS numbers to simulate}
\usage{
getCASnums(func, species, model, defaulttohuman)
}
\arguments{
\item{func}{The user-selected desired output}

\item{species}{The user-selected species}

\item{model}{The user-selected species}

\item{defaulttohuman}{The user-selected human in vitro data preference if an
animal species is selected}
}
\value{
A vector of CASRNs for compounds that are available to simulate in ToCS
}
\description{
This function creates a vector of CASRNs that are available for the user-selected
output, species, model, and human in vitro data preference.
}
\examples{
getCASnums("Concentration-time profiles","Human","pbtk",TRUE)
getCASnums("Steady state concentrations","Rat","1compartment",FALSE)

}
\seealso{
\code{\link[=PreloadCompsInput]{PreloadCompsInput()}}, which calls the current function
}
