% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{getPiped}
\alias{getPiped}
\title{Generate the final list of available compounds available for simulation}
\usage{
getPiped(CASnums, honda, comptype)
}
\arguments{
\item{CASnums}{The CASRN numbers of compounds that have enough data in httk
to simulate for the user's selected parameters}

\item{honda}{The selected IVIVE assumption (either NULL, Honda1, Honda2,
Honda3, or Honda4), if applicable}

\item{comptype}{The user-selected subset of compounds to search; either "Choose
from all available chemicals" or "Choose from only food relevant chemicals"}
}
\value{
A vector of CASRN and compound name pairings for compounds with enough
available data to simulate in httk
}
\description{
This function narrows down the list of compounds from the getCASnums function
by taking into account any IVIVE conditions that the user selected or if the
user is only interested in a list of food relevant chemicals.
}
\seealso{
\code{\link[=PreloadCompsInput]{PreloadCompsInput()}}, which calls the current function
}
