% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{GP_Instructions}
\alias{GP_Instructions}
\title{User interface for the 'Instructions' card on the 'General Parameters' tab}
\usage{
GP_Instructions()
}
\value{
Text with instructions for the user on how to use the app, where to get
help, and where to report bugs
}
\description{
This function outputs the user interface for the 'Instructions' card on the
'General Parameters' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=Instructions_GenPars]{Instructions_GenPars()}},
which is called by the current function
}
