% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{GP_Species}
\alias{GP_Species}
\title{User interface for the 'Species' card on the 'General Parameters' tab}
\usage{
GP_Species()
}
\value{
The 'Species' card with species selection options
}
\description{
User interface for the 'Species' card on the 'General Parameters' tab
}
