% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IndirectFoodAdditives}
\alias{IndirectFoodAdditives}
\title{List of Indirect Food Additives}
\format{
\subsection{\code{IndirectFoodAdditives}}{

A data frame with 3647 rows and 51 columns:
\describe{
\item{CAS Reg No (or other ID)}{Chemical Abstract Service (CAS) Registry Number
for the substance or a numerical code assigned by the Human Foods Program to
those substances that do not have a CAS Registry Number (977nnn-nn-n series).}
\item{Substance}{The name of the ingredient as recognized by the FDA's Human
Foods Program}
\item{Reg col01, Reg col02, Reg col03, Reg col04, Reg col05, Reg col06}{Color
additive regulations (21 CFR Parts 73, 74, 81, 82).}
\item{Reg 01, Reg 02, Reg 03, Reg 04, Reg 05, Reg 06,
Reg 07, Reg 08, Reg 09, Reg 10, Reg 11, Reg 12,Reg 13,
Reg 14, Reg 16, Reg 17, Reg 18, Reg 19, Reg 20, Reg 21}{Food additive
and GRAS regulations (21 CFR Parts 170-186).}
\item{Reg prohibited189}{Prohibited substance regulation (21 CFR Part 189).}
\item{regs Labeling-Standards}{Food labeling and standards regulations
(21 CFR Parts 100-169).}
\item{SYN01, SYN02, SYN03,SYN04, SYN05, SYN06,SYN07, SYN08, SYN09,SYN10,
SYN12, SYN13,SYN14, SYN15, SYN16,SYN17, SYN18, SYN19}{Synonyms for the
substance name.}
...
}
}
}
\source{
\url{http://www.hfpappexternal.fda.gov/scripts/fdcc/?set=IndirectAdditives}
}
\usage{
IndirectFoodAdditives
}
\description{
A list of compounds including their purposes and regulations in the U.S. FDA's
Inventory of Food Contact Substances database
}
\keyword{datasets}
