% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{InitVals_Par}
\alias{InitVals_Par}
\title{Create a vector of initial conditions}
\usage{
InitVals_Par(model, ICopts, pars)
}
\arguments{
\item{model}{The model the user selects}

\item{ICopts}{The user's selection on whether they want to use initial
conditions other than the default, which is zero for all model states}

\item{pars}{A list of parameters to pass into modules}
}
\value{
A named vector of initial conditions for a specific model
}
\description{
This function creates a named vector of initial conditions for the selected
model in concentration-time profile simulations. The names correspond to model
compartments.
}
\seealso{
\code{\link[=UpdatePars]{UpdatePars()}}, which calls the current function
}
