% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{InputRules_Children}
\alias{InputRules_Children}
\title{Generate input validation rules for each individual module}
\usage{
InputRules_Children(
  iv_common,
  iv_adme,
  iv_ss,
  iv_ivive,
  iv_pc,
  input,
  ic_names
)
}
\arguments{
\item{iv_common}{A new shiny input validator object for all common inputs
across modules (function, species, default to human, model, in silico
parameters, preloaded compounds, and uploaded compounds)}

\item{iv_adme}{A new shiny input validator object for all concentration-time
profile specific parameters that other modules do not have}

\item{iv_ss}{A new shiny input validator object for all steady state
concentrations specific parameters that other modules do not have}

\item{iv_ivive}{A new shiny input validator object for all IVIVE specific
parameters that other modules do not have}

\item{iv_pc}{A new shiny input validator object for all parameter calculations
specific parameters that other modules do not have}

\item{input}{The 'input' object used as the prefix for all ID's (should be
entered as 'input')}

\item{ic_names}{A named vector of all initial condition compartment names.
See global.R and input_functions.R for more information.}
}
\value{
Validation rules attached to each child input validator
}
\description{
Generate input validation rules for each individual module
}
\seealso{
\code{\link[=app_server]{app_server()}}, which calls the current function
}
