% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{InputRules_Parents}
\alias{InputRules_Parents}
\title{Combine common validation rules and individual module validation rules into a
single comprehensive input validator for each module}
\usage{
InputRules_Parents(
  parent_adme_iv,
  iv_adme,
  parent_ss_iv,
  iv_ss,
  parent_ivive_iv,
  iv_ivive,
  parent_pc_iv,
  iv_pc,
  iv_common
)
}
\arguments{
\item{parent_adme_iv}{A new shiny input validator object for all concentration-
time profile module inputs}

\item{iv_adme}{A shiny input validator object for all concentration-time
profile specific parameters that other modules do not have}

\item{parent_ss_iv}{A new shiny input validator object for all steady state
concentration module inputs}

\item{iv_ss}{A shiny input validator object for all steady state
concentrations specific parameters that other modules do not have}

\item{parent_ivive_iv}{A new shiny input validator object for all IVIVE
module inputs}

\item{iv_ivive}{A shiny input validator object for all IVIVE specific
parameters that other modules do not have}

\item{parent_pc_iv}{A new shiny input validator object for all parameter
calculations module inputs}

\item{iv_pc}{A shiny input validator object for all parameter calculations
specific parameters that other modules do not have}

\item{iv_common}{A shiny input validator object for all common inputs
across modules (function, species, default to human, model, in silico
parameters, preloaded compounds, and uploaded compounds)}
}
\value{
Validation rules attached to the four parent input validators
}
\description{
Combine common validation rules and individual module validation rules into a
single comprehensive input validator for each module
}
