% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{Instructions_CompSelect_Part1}
\alias{Instructions_CompSelect_Part1}
\title{Generate text to notify user that they must select or upload at least one
compound to simulate}
\usage{
Instructions_CompSelect_Part1()
}
\value{
Text notifying the user of compound requirements
}
\description{
Generate text to notify user that they must select or upload at least one
compound to simulate
}
\seealso{
\code{\link[=CS_Instructions]{CS_Instructions()}}, which calls the current function
}
