% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{Instructions_CompSelect_Part2}
\alias{Instructions_CompSelect_Part2}
\title{Generate text to notify the user of where to see sample uploaded compound data}
\usage{
Instructions_CompSelect_Part2()
}
\value{
Text notifying the user of uploaded compound data instructions
}
\description{
Generate text to notify the user of where to see sample uploaded compound data
}
\seealso{
\code{\link[=CS_Instructions]{CS_Instructions()}}, which calls the current function
}
