% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{Instructions_GenPars}
\alias{Instructions_GenPars}
\title{Generate text for the instructions card under the 'General Parameters' tab}
\usage{
Instructions_GenPars()
}
\value{
Text describing the major uses of ToCS, where to find additional
resources, and where to report bugs
}
\description{
Generate text for the instructions card under the 'General Parameters' tab
}
\examples{
Instructions_GenPars()

}
