% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{Instructions_Reset}
\alias{Instructions_Reset}
\title{Generate text for the 'Reset Session' button}
\usage{
Instructions_Reset()
}
\value{
Text notifying the user of the 'Reset Session' button
}
\description{
Generate text for the 'Reset Session' button
}
\examples{
Instructions_Reset()

}
