% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{Instructions_RunSim}
\alias{Instructions_RunSim}
\title{Generate text for the 'Run Simulation' button}
\usage{
Instructions_RunSim()
}
\value{
Text notifying the user of the 'Run Simulation' button
}
\description{
Generate text for the 'Run Simulation' button
}
\examples{
Instructions_RunSim()

}
