% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_ConcPlot.R
\name{IVIVE_Plot_server}
\alias{IVIVE_Plot_server}
\title{Server function for the IVIVE dose plot}
\usage{
IVIVE_Plot_server(id, ivive_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in IVIVE_Plot_ui()}

\item{ivive_args}{A Shiny reactive list with the output of IVIVEsol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
Server outputs for the IVIVE plot drop down which includes three elements
}
\description{
#' @description
This function generates the outputs defined in the IVIVE_Plot_ui()
function. This connects the download button with the plot to download,
generates the IVIVE dose plot, and creates the text for the plot caption.
}
\seealso{
\code{\link[=IVIVEplotting]{IVIVEplotting()}} and \code{\link[=IVIVEplot_caption]{IVIVEplot_caption()}}, which are both called by
this function, and \code{\link[=IVIVE_server]{IVIVE_server()}}, which calls the current function
}
