% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_ConcPlot.R
\name{IVIVE_Plot_ui}
\alias{IVIVE_Plot_ui}
\title{User interface function for the IVIVE dose plot}
\usage{
IVIVE_Plot_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in IVIVE_Plot_server()}
}
\value{
User interface for the IVIVE plot drop down with four elements
}
\description{
#' @description
This function outputs the user interface for the IVIVE plot drop down in the
results card under the 'Run Simulation' tab. The interface has three outputs:
a plot download button, a plot with calculated IVIVE doses, and a plot caption.
}
