% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_main.R
\name{IVIVE_server}
\alias{IVIVE_server}
\title{Main server function to set up the 'Results' card under the IVIVE module's
'Run Simulation' tab}
\usage{
IVIVE_server(id, pars, runsim, logscale)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in IVIVE_ui()}

\item{pars}{A list of all user input parameters for the entire app}

\item{runsim}{Action button titled 'Run Simulation' pressed by the user}

\item{logscale}{Checkbox input value indicating if the user wanted the y-axis
of plots to be a log10 scale}
}
\value{
The four server outputs that fill the main IVIVE user interface
function consisting of two plots and two tables of IVIVE results
}
\description{
This function calculates the IVIVE solution and outputs the contents of the
UI objects for the IVIVE 'Results' card under the 'Run Simulation' tab. The
contents of each UI object (drop down) are then filled by four separate
server modules.
}
