% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_ConcTable.R
\name{IVIVE_Table_server}
\alias{IVIVE_Table_server}
\title{Server function for the IVIVE dose table}
\usage{
IVIVE_Table_server(id, ivive_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in IVIVE_Table_ui()}

\item{ivive_args}{A Shiny reactive list with the output of IVIVEsol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
Server outputs for the IVIVE table drop down which includes four elements
}
\description{
This function generates the outputs defined in the IVIVE_Table_ui()
function. This connects the download buttons with the table to download,
fills the data table with content, and creates the text for the table caption.
}
