% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_ConcTable.R
\name{IVIVE_Table_ui}
\alias{IVIVE_Table_ui}
\title{User interface function for the IVIVE dose table}
\usage{
IVIVE_Table_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in IVIVE_Table_server()}
}
\value{
User interface for the IVIVE table drop down with four elements
}
\description{
This function outputs the user interface for the IVIVE table drop down in the
results card under the 'Run Simulation' tab. The interface has four outputs:
a table download button, a simulation parameters download button, a table with
calculated IVIVE doses, and a table caption.
}
\seealso{
\code{\link[=IVIVE_ui]{IVIVE_ui()}}, which calls the current function
}
