% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEmod_main.R
\name{IVIVE_ui}
\alias{IVIVE_ui}
\title{Main user interface function to set up the 'Results' card under the IVIVE
module's 'Run Simulation' tab}
\usage{
IVIVE_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in IVIVE_server()}
}
\value{
Four user interface drop down tabs identifying space for two plots
and two tables of IVIVE results.
}
\description{
This function outputs the structure of the user interface for the IVIVE 'Results'
card under the 'Run Simulation' tab. Four drop down tabs appear: one for an
oral equivalent dose (OED) table, one for an OED plot, one for a bioactivity
exposure ratio (BER) table, and one for a BER plot. The user interface contents
of each drop down is then filled by a separate UI module.
}
