% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{IVIVEplot_caption}
\alias{IVIVEplot_caption}
\title{Generate the caption for the IVIVE plot}
\usage{
IVIVEplot_caption(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A text caption for Figure 1
}
\description{
This function generates the caption for the IVIVE oral equivalent dose plot
output. The caption differs depending on whether a single OED or a vector of
OEDs is returned as well as whether exposure data is available.
}
\seealso{
\code{\link[=IVIVE_Plot_server]{IVIVE_Plot_server()}}, which calls the current function
}
