% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{IVIVEplot_labels}
\alias{IVIVEplot_labels}
\title{Generate the title and y-axis label for the IVIVE oral equivalent dose plot}
\usage{
IVIVEplot_labels(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A list with two elements: a title expression and a y-axis expression
}
\description{
This function generates the title and y-axis expressions for the oral equivalent
dose plot in the IVIVE module. The plot's title is dependent upon the model chosen,
and the y-axis label is dependent upon the output concentration and tissue and
output units selected.
}
