% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{IVIVEplot_logscale}
\alias{IVIVEplot_logscale}
\title{Converts a plot with a linear y-axis to a plot with a log10 y-axis scale}
\usage{
IVIVEplot_logscale(
  plt,
  pars,
  OEDSamples_df,
  combined_df,
  OED_data,
  expdata,
  Q5_OED_df
)
}
\arguments{
\item{plt}{A complete ggplot2 plotting object}

\item{pars}{A list of all user input parameters for the entire app}

\item{OEDSamples_df}{A data frame of oral equivalent dose samples for each
compound; The first list output from PlotPrep_df.}

\item{combined_df}{A data frame with OEDs and exposure estimates; The second
output from OEDPoint_Exposure_Plot.}

\item{OED_data}{A data frame of compound names and their OEDs;
The first output from the IVIVEsol function arranged in ascending order of
the OEDs}

\item{expdata}{A data frame of compound names and their exposure data estimates;
The fifth output of the IVIVEsol function}

\item{Q5_OED_df}{A data frame of compound names and their 5th quantile OED;
The second output of the Plotdf_Prep function}
}
\value{
A plot object of OEDs and exposure data (if applicable) with a log10
y-axis scale
}
\description{
This function converts the y-axis scale on a plot object from a linear to a
log10 y-axis.
}
