% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{IVIVEplotting}
\alias{IVIVEplotting}
\title{Plot the oral equivalent dose (OED) and exposure data (if applicable)}
\usage{
IVIVEplotting(OED_data, BioactiveConc, pars, logscale, expdata)
}
\arguments{
\item{OED_data}{A data frame of chemical names and their OEDs (either one per
chemical or many per chemical if samples were returned); The first output
from the IVIVEsol function}

\item{BioactiveConc}{An input not used}

\item{pars}{A list of all user input parameters for the entire app}

\item{logscale}{Checkbox input value indicating if the user wanted the y-axis
of plots to be a log10 scale}

\item{expdata}{A data frame with chemical names and exposure data estimates}
}
\value{
A ggplot2 plotting object
}
\description{
Plot the oral equivalent dose (OED) and exposure data (if applicable)
}
\seealso{
\code{\link[=IVIVE_Plot_server]{IVIVE_Plot_server()}}, which calls the current function, and
\code{\link[=OEDPoint_Exposure_Plot]{OEDPoint_Exposure_Plot()}}, \code{\link[=OEDPoint_NoExposure_Plot]{OEDPoint_NoExposure_Plot()}}, \code{\link[=Plotdf_Prep]{Plotdf_Prep()}},
\code{\link[=OEDSample_Exposure_Plot]{OEDSample_Exposure_Plot()}}, \code{\link[=OEDSample_NoExposure_Plot]{OEDSample_NoExposure_Plot()}}, and
\code{\link[=IVIVEplot_logscale]{IVIVEplot_logscale()}}, which are called by the current function
}
