% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{IVIVEsol}
\alias{IVIVEsol}
\title{Main function to calculate the oral equivalent dose and bioactivity exposure
ratio for a list of chemicals}
\usage{
IVIVEsol(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A list consisting of the data frame of oral equivalent dose solutions,
the bioactive concentration data frame, a data frame with simulation parameters
and physical-chemical data for chemicals simulated, a data frame with bioactivity
exposure ratios (BER), and a data frame with user-uploaded exposure data
}
\description{
This function is the primary function used to calculate the oral equivalent
dose and bioactivity exposure ratio for a list of chemicals, if applicable.
This function calls other functions that do the actual calculations, but it
returns the final output used to generate the specified outputs in the IVIVE
server functions.
}
