% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{loadInSilicoPars}
\alias{loadInSilicoPars}
\title{Load in silico clint, fup, and caco-2 values}
\usage{
loadInSilicoPars(func, species, model, defaulthuman)
}
\arguments{
\item{func}{The user-selected desired output}

\item{species}{The user-selected species}

\item{model}{The user-selected model}

\item{defaulthuman}{The user-selected human in vitro data preference if an
animal species is selected}
}
\value{
A vector of CAS numbers, each of which has enough data to run a simulation
}
\description{
This function loads in silico-generated clint, fup, and caco-2 values for
compounds in httk's chem.physical_and_invitro.data that are lacking enough
data to run simulations. After loading the in silico data, the list of
compounds is paired down to only those with enough parameter data now that
the in silico data has been loaded.
}
