% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{log10breaks}
\alias{log10breaks}
\title{Generate a vector of 10^x breakpoints for the y-axis of a plot}
\usage{
log10breaks(ydata)
}
\arguments{
\item{ydata}{A vector of y-coordinate data to be plotted}
}
\value{
A vector of y-axis breaks
}
\description{
Generate a vector of 10^x breakpoints for the y-axis of a plot
}
\examples{
log10breaks(ydata = c(0.1,0.004,15,130,0.0005))

}
