% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{Model_Input}
\alias{Model_Input}
\title{Generates a drop down list of possible models to simulate}
\usage{
Model_Input(func, spec)
}
\arguments{
\item{func}{The user-selected output}

\item{spec}{The user-selected species}
}
\value{
A drop down list
}
\description{
Generates a drop down list of possible models to simulate
}
\examples{
Model_Input("Concentration-time profiles","Human")
Model_Input("Parameter calculations","Rat")

}
\seealso{
\code{\link[=app_server]{app_server()}}, which calls the current function
}
