% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{modsol}
\alias{modsol}
\title{Main function for the concentration-time profile output}
\usage{
modsol(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A list with three elements containing an array of the concentration-time
profile solution, a matrix of the toxicokinetic summary, and a data frame of
simulation parameters and physical-chemical data for compounds used in the
simulation
}
\description{
This function is the main function responsible for generating the output for
the concentration-time profiles module. This function calls the Run_ADME_Model
function as well as the TKsummary function to actually calculate the
concentration-time profiles and toxicokinetic summaries, but it then formats
those and other outputs to return to the main module server to distribute.
}
\seealso{
\code{\link[=Run_Model_ADME]{Run_Model_ADME()}}
\code{\link[=SetArraySize]{SetArraySize()}}
\code{\link[=TKsummary]{TKsummary()}}
\code{\link[=AssignArrayNames]{AssignArrayNames()}}
\code{\link[=Rearr_TKSumArray]{Rearr_TKSumArray()}}
\code{\link[=StorePars_ADME]{StorePars_ADME()}}, all of which the current function calls.
\code{\link[=ADME_server]{ADME_server()}}, which calls the current function.
}
