% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{MS_Dosing}
\alias{MS_Dosing}
\title{User interface for the 'Dosing' card on the 'Model Specifications' tab}
\usage{
MS_Dosing()
}
\value{
The 'Dosing' card with dosing selection options depending on the
module function chosen
}
\description{
This function outputs the user interface for the 'Dosing' card on the
'Model Specifications' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=selectInput_DoseRoute]{selectInput_DoseRoute()}},
\code{\link[=selectInput_ADMEdoseunits]{selectInput_ADMEdoseunits()}}, \code{\link[=selectInput_NumDoses]{selectInput_NumDoses()}}, \code{\link[=numericInput_InitialDose]{numericInput_InitialDose()}},
\code{\link[=selectInput_MultipleDosesQ]{selectInput_MultipleDosesQ()}}, \code{\link[=numericInput_MultiDoseAmount]{numericInput_MultiDoseAmount()}},
\code{\link[=sliderInput_MultiDoseTime]{sliderInput_MultiDoseTime()}}, \code{\link[=textInput_DoseMatrix]{textInput_DoseMatrix()}}, and
\code{\link[=numericInput_DailyDose]{numericInput_DailyDose()}}, which are called by the current function
}
