% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{multdose_odd}
\alias{multdose_odd}
\title{Has the user entered dosing amounts and times?}
\usage{
multdose_odd(
  value,
  input,
  message = "The dosing administration amounts and times must be entered"
)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}

\item{message}{A character string to return to the user if condition is true}
}
\value{
If condition if FALSE, then return the message. If the condition is
TRUE, return nothing.
}
\description{
Has the user entered dosing amounts and times?
}
