% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{multdose_Select}
\alias{multdose_Select}
\title{Has the user selected a dosing frequency?}
\usage{
multdose_Select(
  value,
  input,
  message = "The dosing frequency must be selected"
)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}

\item{message}{A character string to return to the user if condition is true}
}
\value{
If condition if FALSE, then return the message. If the condition is
TRUE, return nothing.
}
\description{
This function relays to the user that a dosing frequency must be chosen.
}
\seealso{
\code{\link[=InputRules_Children]{InputRules_Children()}}, which calls the current function
}
