% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{not_null}
\alias{not_null}
\title{Has the user selected and/or uploaded compounds to simulate?}
\usage{
not_null(
  value,
  input,
  message = "At least one compound must be selected or uploaded"
)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}

\item{message}{A character string to return to the user if condition is true}
}
\value{
If condition if FALSE, then return the message. If the condition is
TRUE, return nothing.
}
\description{
Has the user selected and/or uploaded compounds to simulate?
}
