% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{Notify_Computing}
\alias{Notify_Computing}
\title{Notify the user that the solution is being computed after the "Run Simulation"
button is clicked}
\usage{
Notify_Computing()
}
\value{
A pop-up notification to the user
}
\description{
Notify the user that the solution is being computed after the "Run Simulation"
button is clicked
}
\examples{
Notify_Computing()

}
