% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_Alpha}
\alias{numericInput_Alpha}
\title{Generates a numeric input box where the user selects their preference on the
ratio of distribution coefficient for totally charges species and that of the
neutral form}
\usage{
numericInput_Alpha(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box where the user selects their preference on the
ratio of distribution coefficient for totally charges species and that of the
neutral form
}
\examples{
numericInput_Alpha("AlphaPar")

}
