% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_CacoDefault}
\alias{numericInput_CacoDefault}
\title{Generates a numeric input box where the user can customize the caco-2 apical-to-basal
membrane permeability value if desired}
\usage{
numericInput_CacoDefault(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box where the user can customize the caco-2 apical-to-basal
membrane permeability value if desired
}
\examples{
numericInput_CacoDefault("caco2default")

}
