% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_ClintPval}
\alias{numericInput_ClintPval}
\title{Generates a numeric input box where the user specifies their preference for
the hepatic clearance p-value threshold}
\usage{
numericInput_ClintPval(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box where the user specifies their preference for
the hepatic clearance p-value threshold
}
\examples{
numericInput_ClintPval("Clint_Pval")

}
