% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_DailyDose}
\alias{numericInput_DailyDose}
\title{Generate a numeric input box for users to enter the total daily dose administered
during steady state concentration simulations}
\usage{
numericInput_DailyDose(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generate a numeric input box for users to enter the total daily dose administered
during steady state concentration simulations
}
\examples{
numericInput_DailyDose("dailydose")

}
