% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_FSBf}
\alias{numericInput_FSBf}
\title{Generates a numeric input box for the user to customize the fraction of fetal
bovine serum used in the in vitro assay for bioactivity data uploaded in in
vitro in vivo extrapolation (IVIVE) simulations}
\usage{
numericInput_FSBf(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box for the user to customize the fraction of fetal
bovine serum used in the in vitro assay for bioactivity data uploaded in in
vitro in vivo extrapolation (IVIVE) simulations
}
\examples{
numericInput_FSBf("FSBf")

}
