% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_ICvalue}
\alias{numericInput_ICvalue}
\title{Generates a numeric input box for the user to specify the initial condition of
a particular model compartment during concentration-time profile simulations}
\usage{
numericInput_ICvalue(id, compartment)
}
\arguments{
\item{id}{Shiny identifier name}

\item{compartment}{Model compartment name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box for the user to specify the initial condition of
a particular model compartment during concentration-time profile simulations
}
\examples{
numericInput_ICvalue("One_gut","gut")
numericInput_ICvalue("Three_AUC","AUC")
numericInput_ICvalue("P_liver","liver")

}
