% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_InitialDose}
\alias{numericInput_InitialDose}
\title{Generate a numeric input box that has users select the amount of initial dose
administered}
\usage{
numericInput_InitialDose(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generate a numeric input box that has users select the amount of initial dose
administered
}
\examples{
numericInput_InitialDose("initdose")

}
\seealso{
\code{\link[=MS_Dosing]{MS_Dosing()}}, which calls the current function
}
