% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_MinFub}
\alias{numericInput_MinFub}
\title{Generates a numeric input box for the user to specify the minimum acceptable
fraction unbound in plasma.}
\usage{
numericInput_MinFub(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box for the user to specify the minimum acceptable
fraction unbound in plasma.
}
\examples{
numericInput_MinFub("min_fub")

}
