% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_MultiDoseAmount}
\alias{numericInput_MultiDoseAmount}
\title{Generate a numeric input box for users to enter the uniform amount of compound
administered during each dose of a multiple dose administration regimen for
ADME simulations}
\usage{
numericInput_MultiDoseAmount(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generate a numeric input box for users to enter the uniform amount of compound
administered during each dose of a multiple dose administration regimen for
ADME simulations
}
\examples{
numericInput_MultiDoseAmount("mult_doseamount")

}
