% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_Quantile}
\alias{numericInput_Quantile}
\title{Generates a numeric input box for users to specify the steady state concentration
quantile used in the in vitro in vivo extrapolation (IVIVE) simulations}
\usage{
numericInput_Quantile(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box for users to specify the steady state concentration
quantile used in the in vitro in vivo extrapolation (IVIVE) simulations
}
\examples{
numericInput_Quantile("quantile")

}
