% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_Samples}
\alias{numericInput_Samples}
\title{Generates a numeric input box for the user to enter the number of Monte Carlo
samples to generate during in vitro in vivo extrapolation (IVIVE) simulations}
\usage{
numericInput_Samples(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box for the user to enter the number of Monte Carlo
samples to generate during in vitro in vivo extrapolation (IVIVE) simulations
}
\examples{
numericInput_Samples("samples")

}
