% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_SimTime}
\alias{numericInput_SimTime}
\title{Generates a numeric input box where the user enters the simulation time for
concentration-time profile simulations}
\usage{
numericInput_SimTime(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box where the user enters the simulation time for
concentration-time profile simulations
}
\examples{
numericInput_SimTime("simtime")
}
