% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{numericInput_SolSteps}
\alias{numericInput_SolSteps}
\title{Generates a numeric input box for the user to specify the number of solver
steps for the ODE solver in the concentration-time profile simulation}
\usage{
numericInput_SolSteps(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A numeric input box
}
\description{
Generates a numeric input box for the user to specify the number of solver
steps for the ODE solver in the concentration-time profile simulation
}
\examples{
numericInput_SolSteps("solversteps")

}
\seealso{
\code{\link[=AP_ModelSolver]{AP_ModelSolver()}}, which calls the current function
}
