% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{OEDPoint_Exposure_Plot}
\alias{OEDPoint_Exposure_Plot}
\title{Plot the oral equivalent dose (OED) for each chemical when the user uploads
exposure data}
\usage{
OEDPoint_Exposure_Plot(OED_data, expdata, y_exp, title_exp)
}
\arguments{
\item{OED_data}{A data frame with chemical names and their OEDs}

\item{expdata}{A data frame with chemical names and their exposure estimates}

\item{y_exp}{The plot's y-axis label expression}

\item{title_exp}{The plot's title expression}
}
\value{
A list with two elements: a ggplot2 plotting object and a data frame
with both OED and exposure data
}
\description{
Plot the oral equivalent dose (OED) for each chemical when the user uploads
exposure data
}
