% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{OEDPoint_NoExposure_Plot}
\alias{OEDPoint_NoExposure_Plot}
\title{Plot the oral equivalent dose (OED) for each chemical when there is no exposure
data uploaded}
\usage{
OEDPoint_NoExposure_Plot(OED_data, y_exp, title_exp)
}
\arguments{
\item{OED_data}{A data frame with chemical names and their OEDs}

\item{y_exp}{The plot's y-axis label expression}

\item{title_exp}{The plot's title expression}
}
\value{
A ggplot2 plotting object
}
\description{
Plot the oral equivalent dose (OED) for each chemical when there is no exposure
data uploaded
}
\seealso{
\code{\link[=IVIVEplotting]{IVIVEplotting()}}, which calls the current function
}
