% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{OEDSample_Exposure_Plot}
\alias{OEDSample_Exposure_Plot}
\title{Plot the oral equivalent dose samples and user-uploaded exposure data estimates}
\usage{
OEDSample_Exposure_Plot(OEDSamples_df, Q5_OED_df, expdata, y_exp, title_exp)
}
\arguments{
\item{OEDSamples_df}{A data frame with chemical names and their OED samples}

\item{Q5_OED_df}{A data frame with chemical names and their 5th OED quantile value}

\item{expdata}{A data frame with chemical names and their exposure estimates}

\item{y_exp}{The plot's y-axis label expression}

\item{title_exp}{The plot's title expression}
}
\value{
A list with two elements: a ggplot2 plotting object and a data frame
with exposure data
}
\description{
This function generates the OED samples ggplot object no exposure data is
uploaded by the user.
}
\seealso{
\code{\link[=FillExposureData]{FillExposureData()}}, which is called by the current function, and
\code{\link[=IVIVEplotting]{IVIVEplotting()}}, which calls the current function
}
