% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{OEDSample_NoExposure_Plot}
\alias{OEDSample_NoExposure_Plot}
\title{Plot the oral equivalent dose samples when the user does not upload exposure data}
\usage{
OEDSample_NoExposure_Plot(OEDSamples_df, Q5_OED_df, y_exp, title_exp)
}
\arguments{
\item{OEDSamples_df}{A data frame with chemical names and their OED samples}

\item{Q5_OED_df}{A data frame with chemical names and their 5th OED quantile value}

\item{y_exp}{The plot's y-axis label expression}

\item{title_exp}{The plot's title expression}
}
\value{
A ggplot2 plotting object
}
\description{
This function generates the OED samples ggplot object when no exposure data is
uploaded by the user.
}
\seealso{
\code{\link[=IVIVEplotting]{IVIVEplotting()}}, which calls the current function
}
