% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{OutputTimes_Par}
\alias{OutputTimes_Par}
\title{Generate a vector specifying times to output the time course solution in the
ADME module simulation}
\usage{
OutputTimes_Par(pars)
}
\arguments{
\item{pars}{A list of parameters to pass into modules}
}
\value{
A vector of times to output the ADME concentration-time profile solution
}
\description{
Generate a vector specifying times to output the time course solution in the
ADME module simulation
}
