% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{ParNames}
\alias{ParNames}
\title{Compile the names of all user-selected parameters across all modules}
\usage{
ParNames()
}
\value{
A list of parameter id names
}
\description{
Compile the names of all user-selected parameters across all modules
}
\seealso{
\code{\link[=GatherInputVars]{GatherInputVars()}}, which calls the current function
}
