% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCmod_EVPlot.R
\name{PC_EVPlot_server}
\alias{PC_EVPlot_server}
\title{Server function for the parameter calculations parameter plot}
\usage{
PC_EVPlot_server(id, pc_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in PC_EVPlot_ui()}

\item{pc_args}{A Shiny reactive list with the output of Parsol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
Server outputs for the parameter plot drop down which includes three
elements
}
\description{
This function contains the output elements for the parameter calculations
parameter plot. This server contains three output elements: a
download button, a plot, and a plot caption.
}
\seealso{
\code{\link[=plotPar]{plotPar()}}, which is called by the current function, and \code{\link[=PC_server]{PC_server()}},
which calls the current function
}
