% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCmod_EVPlot.R
\name{PC_EVPlot_ui}
\alias{PC_EVPlot_ui}
\title{User interface function for the parameter calculations parameter plot}
\usage{
PC_EVPlot_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in PC_EVPlot_server()}
}
\value{
User interface for the parameter plot drop down with three elements
}
\description{
This function outputs the structure of the user interface for the Parameter
Calculations 'Results' card under the 'Run Simulation' tab. Three drop down
tabs appear: a download plot button, a plot of parameter values, and a plot
caption. The user interface contents of each drop down is then filled by a
separate UI module.
}
