% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCmod_Pars.R
\name{PC_Pars_ui}
\alias{PC_Pars_ui}
\title{User interface function for the paramater calculations simulation parameters
download option}
\usage{
PC_Pars_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in PC_Pars_server()}
}
\value{
User interface for the parameter download drop down with one element
}
\description{
User interface function for the paramater calculations simulation parameters
download option
}
