% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCmod_PartitionCoefficientsTable.R
\name{PC_PCTable_server}
\alias{PC_PCTable_server}
\title{Server function for the parameter calculations partition coefficients table}
\usage{
PC_PCTable_server(id, pc_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in PC_PCTable_ui()}

\item{pc_args}{A Shiny reactive list with the output of Parsol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
Server outputs for the partition coefficients table with three elements
}
\description{
Server function for the parameter calculations partition coefficients table
}
\seealso{
\code{\link[=PC_server]{PC_server()}}, which calls the current function
}
