% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCmod_main.R
\name{PC_server}
\alias{PC_server}
\title{Main server function to set up the 'Results' card under the Parameter
Calculation module's Run Simulation' tab}
\usage{
PC_server(id, pars, runsim, logscale)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in PC_ui()}

\item{pars}{A list of all user input parameters for the entire app}

\item{runsim}{Action button titled 'Run Simulation' pressed by the user}

\item{logscale}{Checkbox input value indicating if the user wanted the y-axis
of plots to be a log10 scale}
}
\value{
The five server outputs that fill the main Parameter Calculations
user interface function consisting of two plots, two tables, and a download
button of simulation parameters of Parameter Calculation results
}
\description{
This function acts as the main function to fill the user interface for the
parameter calculations module. The server outputs include five elements: two
plots, two tables, and a download button. Each server element has its own
server function.
}
\seealso{
\code{\link[=validate_text_PC]{validate_text_PC()}}, \code{\link[=Parsol]{Parsol()}}, \code{\link[=PC_EVPlot_server]{PC_EVPlot_server()}}, \code{\link[=PC_EVTable_server]{PC_EVTable_server()}},
\code{\link[=PC_PCPlot_server]{PC_PCPlot_server()}}, \code{\link[=PC_PCTable_server]{PC_PCTable_server()}}, and \code{\link[=PC_Pars_server]{PC_Pars_server()}}, which
are all called by this function, and \code{\link[=Run_Simulation]{Run_Simulation()}}, which calls the
current function
}
