% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCmod_main.R
\name{PC_ui}
\alias{PC_ui}
\title{Main user interface function to set up the 'Results' card under the Parameter
Calculations module's 'Run Simulation' tab}
\usage{
PC_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in PC_server()}
}
\value{
Five user interface drop down tabs identifying space for two plots,
two tables, and a download button of parameter calculation results.
}
\description{
#' @description
This function outputs the structure of the user interface for the Parameter
Calculations 'Results' card under the 'Run Simulation' tab. Five drop down
tabs appear: one for a parameter plot, one for a parameter table, one for a
partition coefficient plot, one for a partition coefficient table, and one for
a download option for simulation parameters. The user interface contents of
each drop down is then filled by a separate UI module.
}
