% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_outputfuncs.R
\name{plot_labels}
\alias{plot_labels}
\title{Steady state module plot labeling function}
\usage{
plot_labels(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A list containing plots labels for the title and y-axis.
}
\description{
Steady state module plot labeling function
}
