% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{plot_logscale}
\alias{plot_logscale}
\title{Generate a plot with a log10 scale y-axis}
\usage{
plot_logscale(plt, sol_vec)
}
\arguments{
\item{plt}{ggplot2 plot object to add the log10 y-axis to}

\item{sol_vec}{A vector of y-values to be plotted on plt}
}
\value{
A plot that is plt with a log10 y-axis
}
\description{
This function transforms a given plot so that its y-axis scale is a log10 scale
instead of a linear scale.
}
\seealso{
\code{\link[=scat_plot]{scat_plot()}}, \code{\link[=plotPar]{plotPar()}}, and \code{\link[=plotPCs]{plotPCs()}}, which calls the current
function, and \code{\link[=log10breaks]{log10breaks()}}, which the current function calls
}
