% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{Plotdf_Prep}
\alias{Plotdf_Prep}
\title{Rearrange the data frame of compounds and their OED samples by median OED
sample and separate the 5th quantile OED for each chemical}
\usage{
Plotdf_Prep(df, pars)
}
\arguments{
\item{df}{A data frame with chemical names and outputted OED samples from
IVIVEsol()}

\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A list with two elements: the OED samples and the 5th quantile OEDs
arranged by median value
}
\description{
This function prepares the data frames for plotting the OED when all OED samples
are returned instead of a singular OED per compound. The compounds are arranged
in ascending order based on the median OED sample for each compound and then
the data frame with the 5th quantile OED is arranged in that same order.
}
