% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_outputfuncs.R
\name{plotPar}
\alias{plotPar}
\title{Generate plots of calculated elimination rate, volume of distribution, half-life,
and total plasma clearance values for a list of chemicals}
\usage{
plotPar(soldata, pars, logscale)
}
\arguments{
\item{soldata}{A data frame with calculated elimination rates, volumes of
distribution, half life, and total clearance measures for each simulated
compound.}

\item{pars}{A list of all user input parameters for the entire app}

\item{logscale}{Checkbox input value indicating if the user wanted the y-axis
of plots to be a log10 scale; either 'TRUE' for log10 y-axis or 'FALSE' for a
linear y-axis}
}
\value{
A list of four ggplot2 objects for plots of elimination rate, volume
of distribution, half life, and total plasma clearance
}
\description{
This function generates plots for calculated elimination rate, volume of
distribution, half-life, and total plasma clearance values for a list of
compounds. Each plot has its values arranged in ascending order and plot on a
log10 y-axis scale, if desired by the user.
}
\seealso{
\code{\link[=PC_EVPlot_server]{PC_EVPlot_server()}}, which calls the current function, and \code{\link[=plot_logscale]{plot_logscale()}},
which the current function calls
}
