% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_outputfuncs.R
\name{plotPCs}
\alias{plotPCs}
\title{Generate partition coefficient plots for simulated chemicals}
\usage{
plotPCs(soldata, pars, logscale)
}
\arguments{
\item{soldata}{A data frame with calculated partition coefficients for each
simulated compound.}

\item{pars}{A list of all user input parameters for the entire app}

\item{logscale}{Checkbox input value indicating if the user wanted the y-axis
of plots to be a log10 scale; either 'TRUE' for log10 y-axis or 'FALSE' for a
linear y-axis}
}
\value{
A list of four ggplot2 objects for plots of elimination rate, volume
of distribution, half life, and total plasma clearance
}
\description{
This function generates plots of partition coefficients for all simulated
compounds. Each plot represents a different tissue and that tissue's partition
coefficients are arranged in ascending order. The plots are plotted on a log10
y-axis scale, if desired by the user.
}
\seealso{
\code{\link[=PC_PCPlot_server]{PC_PCPlot_server()}}, which calls the current function, and
\code{\link[=plot_logscale]{plot_logscale()}}, which the current function calls
}
