% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{plottingfunc_all}
\alias{plottingfunc_all}
\title{Generate the ggplot2 plotting object for the multi-curve plot}
\usage{
plottingfunc_all(sol_array)
}
\arguments{
\item{sol_array}{The concentration-time profile solution array where entries
signify the concentration of a compound at a specific time in a particular
compartment. Pages of the array correspond to compounds, rows to time points,
and columns to model compartments and outputs.}
}
\value{
A list containing two elements, the first one being a plot
list with each element of the list representing a different compound (no plot
legends) and the second element being a data frame of hexadecimal RGB
triplet colors that defines the plotting color for each compound
}
\description{
Generate the ggplot2 plotting object for the multi-curve plot
}
