% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{PreloadComps_UI}
\alias{PreloadComps_UI}
\title{Compile outputs for the 'Preloaded Compounds' card under the 'Compound Selection'
card}
\usage{
PreloadComps_UI(func, spec, defaulthuman, model, insilicopars, honda, comptype)
}
\arguments{
\item{func}{User selection with the input$func ID; either "Concentration-time
profiles", "Steady state concentrations", "In vitro in vivo extrapolation
(IVIVE)", and "Parameter calculations"}

\item{spec}{User selection with the input$spec ID; either "Dog","Human","Mouse",
"Rabbit","Rat"}

\item{defaulthuman}{User selection with input$defaulttoHuman; either "Yes" for
using human in vitro data in place of animal data or "No" for only using
animal data}

\item{model}{User selection with the input$model ID; either "3compartmentss",
"Schmitt","1compartment","3compartment", "pbtk" or "fetal_pbtk"}

\item{insilicopars}{User selection with the input$insilicopars ID; either
"Yes, load in silico parameters" or "No, do not load in silico parameters"}

\item{honda}{User selection with the input$HondaIVIVE ID; either "NULL","Honda1",
"Honda2","Honda3","Honda4"}

\item{comptype}{User selection with the input$comptype ID; either "Choose from
all available chemicals" or "Choose from only food relevant chemicals"}
}
\value{
The user interface objects (drop down selections) for the 'Preloaded
Compounds' card on the 'Compound Selection' tab. The objects outputted vary
depending on the user's input selection.
}
\description{
Compile outputs for the 'Preloaded Compounds' card under the 'Compound Selection'
card
}
\seealso{
\code{\link[=app_server]{app_server()}}, which calls the current function
}
