% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{PrepExposureData}
\alias{PrepExposureData}
\title{Prepare the exposure data file for plotting}
\usage{
PrepExposureData(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A data frame with exposure data
}
\description{
This function prepares the user uploaded exposure data file for plotting. The
function assures that the chemicals in the exposure data file are in the same
order as the original chemical list, updates the data if the model output units
are different than the default of mg/kg/day, and calculates the maximal exposure
data estimate for each compound. The updated exposure data frame is returned.
}
