% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{Rearr_TKSumArray}
\alias{Rearr_TKSumArray}
\title{Rearrange the toxicokinetic (TK) array into a TK matrix}
\usage{
Rearr_TKSumArray(tk_sum_array, pars)
}
\arguments{
\item{tk_sum_array}{The array of TK summary data where each page of the array
represents the TK summary of a particular compound with rows representing
each model compartment and columns being Tmax, Cmax, and AUC}

\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A matrix with rows representing model compartments and 3n columns (where
n is the compounds of simulated compounds) with each compound having a Tmax,
Cmax, and AUC column
}
\description{
Rearrange the toxicokinetic (TK) array into a TK matrix
}
