% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{RemoveCols}
\alias{RemoveCols}
\title{Remove columns from the concentration-time profile solution matrix}
\usage{
RemoveCols(sol, model)
}
\arguments{
\item{sol}{The solution matrix from httk's 'solve_model' function}

\item{model}{The simulated model; either "1compartment","3compartment","pbtk",
"fetal_pbtk"}
}
\value{
The trimmed solution matrix with less columns than sol; All columns
after the AUC column are removed if the model is not "fetal_pbtk" and if the
model is "fetal_pbtk", then all columns after Qthyroid are removed.
}
\description{
This function cuts off columns that are considered 'extra' from the
concentration-time profile solution matrix outputted from httk's solve_model
function. This includes columns that are essentially 'repeat' quantities (i.e.
removes a column specifying the amount of a chemical in compartment i when
the concentration of the chemical in compartment i is already outputted).
}
\seealso{
\code{\link[=Run_ADME_model]{Run_ADME_model()}}, which calls the current function
}
