% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{returntimes_cond}
\alias{returntimes_cond}
\title{Has the user entered a beginning output time of 91 days or later?}
\usage{
returntimes_cond(
  value,
  input,
  message = "A beginning output time of 91 days (13 weeks) or later must be entered"
)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}

\item{message}{A character string to return to the user if condition is true}
}
\value{
If condition if FALSE, then return the message. If the condition is
TRUE, return nothing.
}
\description{
This function relays to the user that return times for the fetal_pbtk model
must start at 91 days or later. This is a restriction of the model that it
only outputs times at 13 weeks or later.
}
\seealso{
\code{\link[=InputRules_Children]{InputRules_Children()}}, which calls the current function
}
