% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{RS_Actions}
\alias{RS_Actions}
\title{User interface for the 'Actions' card on the 'Run Simulation' tab}
\usage{
RS_Actions()
}
\value{
The 'Actions' card which provides instructions to the user on how to
run simulations, two buttons for the user to press to run or reset the
simulation, and the option to log scale plots (if applicable for the module
chosen)
}
\description{
User interface for the 'Actions' card on the 'Run Simulation' tab
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=Instructions_RunSim]{Instructions_RunSim()}},
\code{\link[=Instructions_Reset]{Instructions_Reset()}}, and \code{\link[=checkboxInput_Log]{checkboxInput_Log()}}, which are called by the
current function
}
