% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{RS_Results}
\alias{RS_Results}
\title{User interface for the 'Results' card on the 'Run Simulation' tab}
\usage{
RS_Results()
}
\value{
The 'Results' card which displays a number of drop down menus filled
with output results depending on the module function chosen
}
\description{
User interface for the 'Results' card on the 'Run Simulation' tab
}
