% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{RS_SelectedCompounds}
\alias{RS_SelectedCompounds}
\title{User interface for the 'Selected Compounds' card on the 'Run Simulation' tab}
\usage{
RS_SelectedCompounds()
}
\value{
The 'Selected Compounds' card which shows a table of all compounds
selected by the user for simulation
}
\description{
User interface for the 'Selected Compounds' card on the 'Run Simulation' tab
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function
}
