% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{Run_ADME_Model}
\alias{Run_ADME_Model}
\title{Solve the ADME model}
\usage{
Run_ADME_Model(i, pars)
}
\arguments{
\item{i}{A numerical value; the index number of the compound to simulate from
the user selected chemical list}

\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A matrix with time course data for the selected model and simulation
parameters. The first column has time points and the remaining columns have
concentrations for each model compartment, plasma concentration, and area
under the curve of the plasma compartment. Each row represents a concentration
at the listed time.
}
\description{
This function solves the PBTK model to generate concentration-time profiles for
the list of selected compounds. These concentrations are generated using the
'httk' package's 'solve_model' function for the selected model and parameters.
}
