% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{Run_Simulation}
\alias{Run_Simulation}
\title{Run the selected module with all user selected inputs}
\usage{
Run_Simulation(
  parent_adme_iv,
  parent_ss_iv,
  parent_ivive_iv,
  parent_pc_iv,
  input,
  AllInputs
)
}
\arguments{
\item{parent_adme_iv}{A shiny input validator object for all concentration-
time profile module inputs}

\item{parent_ss_iv}{A shiny input validator object for all steady state
concentration module inputs}

\item{parent_ivive_iv}{A shiny input validator object for all IVIVE
module inputs}

\item{parent_pc_iv}{A shiny input validator object for all parameter
calculations module inputs}

\item{input}{The 'input' object used as the prefix for all ID's (should be
entered as 'input')}

\item{AllInputs}{The named list of all parameter values created in the GUI,
named accordingly in parnames()}
}
\value{
The module outputs for the selected module or an error notification
indicating missing or invalid parameters
}
\description{
Run the selected module with all user selected inputs
}
