% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_outputfuncs.R
\name{scat_plot}
\alias{scat_plot}
\title{Create a scatter plot of steady state concentrations for the steady state
concentrations module}
\usage{
scat_plot(sol, pars, logscale)
}
\arguments{
\item{sol}{A data frame of steady state concentrations}

\item{pars}{A list of all user input parameters for the entire app}

\item{logscale}{The log10 y-axis scale checkbox input to signify whether the
user wants a plot to have a log10 y-axis scale}
}
\value{
A scatter plot object of steady state concentrations, either with a
linear or log10 scale y-axis
}
\description{
This function generates the scatter plot ggplot object containing the steady
state concentrations for all simulated compounds. The plot has a log10 y-axis
scale if desired by the user.
}
\seealso{
\code{\link[=SS_ConcPlot_server]{SS_ConcPlot_server()}}, which calls the current function, and \code{\link[=plot_labels]{plot_labels()}}
and \code{\link[=plot_logscale]{plot_logscale()}}, which are called by the current function
}
