% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_AdjFub}
\alias{selectInput_AdjFub}
\title{Generates a drop down list of chemical fraction unbound in plasma calculation
preferences for the user to select from}
\usage{
selectInput_AdjFub(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list of chemical fraction unbound in plasma calculation
preferences for the user to select from
}
\examples{
selectInput_AdjFub("adj_fub")

}
\seealso{
\code{\link[=AP_ModelConditions]{AP_ModelConditions()}}, which calls the current function
}
