% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_ADMEdoseunits}
\alias{selectInput_ADMEdoseunits}
\title{Generate a drop down list of dose units to select from}
\usage{
selectInput_ADMEdoseunits(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generate a drop down list of dose units to select from
}
\examples{
selectInput_ADMEdoseunits("doseunits")

}
\seealso{
\code{\link[=MS_Dosing]{MS_Dosing()}}, which calls the current function
}
