% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_Bioactive}
\alias{selectInput_Bioactive}
\title{Generates a drop down list of options for the user to select which chemical
concentration is treated as bioactive in vivo}
\usage{
selectInput_Bioactive(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list of options for the user to select which chemical
concentration is treated as bioactive in vivo
}
\examples{
selectInput_Bioactive("bioactiveIVIVE")
}
\seealso{
\code{\link[=AP_ModelConditions]{AP_ModelConditions()}}, which calls the current function
}
