% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_CompPreference}
\alias{selectInput_CompPreference}
\title{Generate the drop down menu where the user can specify the types of chemicals
to choose from}
\usage{
selectInput_CompPreference(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
Drop down menu with options to either set simulated chemicals as a
demo list, select chemicals from the entire available list, or select chemicals
from a list of food or food-related chemicals only
}
\description{
Generate the drop down menu where the user can specify the types of chemicals
to choose from
}
\examples{
selectInput_CompPreference("PreferredCompLst")

}
