% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_DefaultToHuman}
\alias{selectInput_DefaultToHuman}
\title{Generate the 'use of human data in place of animal data if unavailable' drop
down menu}
\usage{
selectInput_DefaultToHuman(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
Drop down menu for the human data preference use in ToCS, if applicable
}
\description{
Generate the 'use of human data in place of animal data if unavailable' drop
down menu
}
\examples{
selectInput_DefaultToHuman("defaulttoHuman")

}
