% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_DoseRoute}
\alias{selectInput_DoseRoute}
\title{Generate a drop down menu for administration method of compounds during ADME
simulation}
\usage{
selectInput_DoseRoute(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop drop list
}
\description{
Generate a drop down menu for administration method of compounds during ADME
simulation
}
\examples{
selectInput_DoseRoute("doseroute")

}
\seealso{
\code{\link[=MS_Dosing]{MS_Dosing()}}, which calls the current function
}
