% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_Fgut}
\alias{selectInput_Fgut}
\title{Generates a drop down list where the user can select their preference for whether
the caco-2 apical-to-basal value defined in numericInput_CacoDefault() should
be used to estimate the in vivo measured fraction of an oral dose that passes
gut metabolism and clearance if bioavailability data is unavailable}
\usage{
selectInput_Fgut(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list where the user can select their preference for whether
the caco-2 apical-to-basal value defined in numericInput_CacoDefault() should
be used to estimate the in vivo measured fraction of an oral dose that passes
gut metabolism and clearance if bioavailability data is unavailable
}
\examples{
selectInput_Fgut("caco_fgut")

}
\seealso{
\code{\link[=AP_Bioavailability]{AP_Bioavailability()}}, which calls the current function
}
