% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_Function}
\alias{selectInput_Function}
\title{Generate the 'Output' drop down menu}
\usage{
selectInput_Function(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
Drop down menu for the desired output of ToCS
}
\description{
Generate the 'Output' drop down menu
}
\examples{
selectInput_Function("func")

}
