% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_HondaCond}
\alias{selectInput_HondaCond}
\title{Generates a drop down list of IVIVE assumptions for the user to select from
during the in vitro in vivo extrapolation (IVIVE) simulations}
\usage{
selectInput_HondaCond(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list of IVIVE assumptions for the user to select from
during the in vitro in vivo extrapolation (IVIVE) simulations
}
\examples{
selectInput_HondaCond("HondaIVIVE")

}
