% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_InitialCondCustom}
\alias{selectInput_InitialCondCustom}
\title{Generates a drop down list of options for whether the user wants to customize
initial condition}
\usage{
selectInput_InitialCondCustom(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list of options for whether the user wants to customize
initial condition
}
\examples{
selectInput_InitialCondCustom("init_cond_opts")

}
\seealso{
\code{\link[=AP_ModelConditions]{AP_ModelConditions()}}, which calls the current function
}
