% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_InSilicoPars}
\alias{selectInput_InSilicoPars}
\title{Generate the 'In silico parameter preference' drop down menu}
\usage{
selectInput_InSilicoPars(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
Drop down menu with options to either use in silico generated clint
(instrinic hepatic clearance) and fup (fraction unbound in plasma) in absence
of in vitro data for the selected species
}
\description{
Generate the 'In silico parameter preference' drop down menu
}
\examples{
selectInput_InSilicoPars("insilicopars")

}
\seealso{
\code{\link[=MS_Model]{MS_Model()}}, which calls the current function
}
