% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_IVIVEoutunits}
\alias{selectInput_IVIVEoutunits}
\title{Generates a drop down list of dose output units for the in vitro in vivo
extrapolation (IVIVE) simulations}
\usage{
selectInput_IVIVEoutunits(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list of dose output units for the in vitro in vivo
extrapolation (IVIVE) simulations
}
\examples{
selectInput_IVIVEoutunits("modelIVIVEout_units")

}
\seealso{
\code{\link[=AP_OutputSpecification]{AP_OutputSpecification()}}, which calls the current function
}
