% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_Keep100}
\alias{selectInput_Keep100}
\title{Generates a drop down list where the user can select their preference for
whether to keep F_abs and F_gut at 100\% availability}
\usage{
selectInput_Keep100(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list where the user can select their preference for
whether to keep F_abs and F_gut at 100\% availability
}
\examples{
selectInput_Keep100("caco_keep100")

}
