% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_MultipleDosesQ}
\alias{selectInput_MultipleDosesQ}
\title{Generate a drop down list of multiple dosing frequency options for ADME
simulations}
\usage{
selectInput_MultipleDosesQ(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generate a drop down list of multiple dosing frequency options for ADME
simulations
}
\examples{
selectInput_MultipleDosesQ("multdose")

}
