% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_ODEmethod}
\alias{selectInput_ODEmethod}
\title{Generates a drop down list of options for the ODE solver method for concentration-time
profile simulations}
\usage{
selectInput_ODEmethod(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list of options for the ODE solver method for concentration-time
profile simulations
}
\examples{
selectInput_ODEmethod("odemethod")

}
\seealso{
\code{\link[=AP_ModelSolver]{AP_ModelSolver()}}, which calls the current function
}
