% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_OutConc}
\alias{selectInput_OutConc}
\title{Generates a drop down list of the output concentration type for the steady state
concentration and IVIVE outputs}
\usage{
selectInput_OutConc(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list of the output concentration type for the steady state
concentration and IVIVE outputs
}
\examples{
selectInput_OutConc("output_concSS")
selectInput_OutConc("output_concIVIVE")

}
\seealso{
\code{\link[=AP_OutputSpecification]{AP_OutputSpecification()}}, which calls the current function
}
