% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_rb2p}
\alias{selectInput_rb2p}
\title{Generates a drop down list where the user selects its ratio of blood to plasma
preferences for the specified model}
\usage{
selectInput_rb2p(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list where the user selects its ratio of blood to plasma
preferences for the specified model
}
\examples{
selectInput_rb2p("rb2p")

}
