% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_Regression}
\alias{selectInput_Regression}
\title{Generates a drop down list where the user chooses their preference on whether
regressions are used in the calculation of partition coefficients}
\usage{
selectInput_Regression(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list where the user chooses their preference on whether
regressions are used in the calculation of partition coefficients
}
\examples{
selectInput_Regression("regression")

}
\seealso{
\code{\link[=AP_ModelConditions]{AP_ModelConditions()}}, which calls the current function
}
