% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_RestrictClear}
\alias{selectInput_RestrictClear}
\title{Generates a drop down list of clearance method preferences for the user to
select from}
\usage{
selectInput_RestrictClear(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list of clearance method preferences for the user to
select from
}
\examples{
selectInput_RestrictClear("restrict_clear")

}
