% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_ReturnSamps}
\alias{selectInput_ReturnSamps}
\title{Generates a drop down list for the user to select whether the want to return
only a single OED estimate or all OED estimates for the in vitro in vivo
extrapolation (IVIVE) simulation}
\usage{
selectInput_ReturnSamps(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list for the user to select whether the want to return
only a single OED estimate or all OED estimates for the in vitro in vivo
extrapolation (IVIVE) simulation
}
\examples{
selectInput_ReturnSamps("returnsamples")

}
