% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_Species}
\alias{selectInput_Species}
\title{Generate the 'Species' drop down menu}
\usage{
selectInput_Species(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
Drop down menu for the desired species to simulate in ToCS
}
\description{
Generate the 'Species' drop down menu
}
\examples{
selectInput_Species("spec")

}
\seealso{
\code{\link[=GP_Species]{GP_Species()}}, which calls the current function
}
