% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_SSoutunits}
\alias{selectInput_SSoutunits}
\title{Generates a drop down list of potential output concentrations for the steady
state concentrations output}
\usage{
selectInput_SSoutunits(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A drop down list
}
\description{
Generates a drop down list of potential output concentrations for the steady
state concentrations output
}
\examples{
selectInput_SSoutunits("modelSSout_units")

}
\seealso{
\code{\link[=AP_OutputSpecification]{AP_OutputSpecification()}}, which calls the current function
}
