% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{selectInput_Tissue}
\alias{selectInput_Tissue}
\title{Generates a drop down list of tissue types for the steady state concentration
and IVIVE outputs}
\usage{
selectInput_Tissue(id)
}
\arguments{
\item{id}{Shiny identifier}
}
\value{
A drop down list
}
\description{
Generates a drop down list of tissue types for the steady state concentration
and IVIVE outputs
}
\examples{
selectInput_Tissue("tissueSS")
selectInput_Tissue("tissueIVIVE")

}
\seealso{
\code{\link[=AP_OutputSpecification]{AP_OutputSpecification()}}, which calls the current function
}
