% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{sliderInput_MultiDoseTime}
\alias{sliderInput_MultiDoseTime}
\title{Generate a slider input bar to select how often a uniform dose of compound
is administered during an ADME simulation}
\usage{
sliderInput_MultiDoseTime(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A slider input
}
\description{
Generate a slider input bar to select how often a uniform dose of compound
is administered during an ADME simulation
}
\examples{
sliderInput_MultiDoseTime("mult_dosetime")

}
