% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{sliderInput_RTol}
\alias{sliderInput_RTol}
\title{Generates a slider input where the user can specify the ODE solver's relative
tolerance for concentration-time profile simulations}
\usage{
sliderInput_RTol(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A slider input
}
\description{
Generates a slider input where the user can specify the ODE solver's relative
tolerance for concentration-time profile simulations
}
\examples{
sliderInput_RTol("rtol")
}
\seealso{
\code{\link[=AP_ModelSolver]{AP_ModelSolver()}}, which calls the current function
}
