% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmod_ConcPlot.R
\name{SS_ConcPlot_server}
\alias{SS_ConcPlot_server}
\title{Server function for the steady state concentrations plot}
\usage{
SS_ConcPlot_server(id, ss_args)
}
\arguments{
\item{id}{Shiny identifier name; must match the identifier in the SS_ConcPlot_ui()}

\item{ss_args}{A Shiny reactive list with the output of SS_sol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
Server function for the steady state concentrations plot which includes
four elements: a plot download button, the steady state concentrations plot,
and a plot caption
}
\description{
This function contains the output elements for the steady state
concentrations plot. This server contains three output elements: a
download button, a plot, and a plot caption.
}
\seealso{
\code{\link[=scat_plot]{scat_plot()}}, which the current function calls, and \code{\link[=SS_server]{SS_server()}},
which calls the current function
}
