% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmod_ConcPlot.R
\name{SS_ConcPlot_ui}
\alias{SS_ConcPlot_ui}
\title{User interface for the steady state concentrations plot}
\usage{
SS_ConcPlot_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must match the identifier in the SS_ConcPlot_server()}
}
\value{
User interface for the steady state concentrations plot which includes
four elements: a plot download button, the steady state concentrations plot,
and a plot caption
}
\description{
User interface for the steady state concentrations plot
}
\seealso{
\code{\link[=SS_ui]{SS_ui()}}, which calls the current function
}
