% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmod_ConcTable.R
\name{SS_ConcTable_server}
\alias{SS_ConcTable_server}
\title{Server function for the steady state concentrations table}
\usage{
SS_ConcTable_server(id, ss_args)
}
\arguments{
\item{id}{Shiny identifier name; must match the identifier in SS_ConcTable_ui()}

\item{ss_args}{A Shiny reactive list with the output of SS_sol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
Server outputs for the steady state concentrations table which includes
four elements: a table download button, a simulation parameters download button,
the steady state concentrations table, and a table caption
}
\description{
This function contains the output elements for the steady state
concentrations table. This server contains four output elements: a
table download button, a simulation download button, a table, and a table
caption.
}
\seealso{
\code{\link[=SS_server]{SS_server()}}, which calls the current function
}
