% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmod_SSDayTable.R
\name{SS_DayTable_server}
\alias{SS_DayTable_server}
\title{The server function for the days to steady state table}
\usage{
SS_DayTable_server(id, ss_args)
}
\arguments{
\item{id}{Shiny identifier name; identifier must match that in SS_DayTable_ui()}

\item{ss_args}{A Shiny reactive list with the output of SS_sol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
The server function for the days to steady state table which includes
three outputs: a table download button, a table, and a table caption
}
\description{
This function contains the output elements for the days to steady state table.
This server contains three output elements: a table download button, a table,
and a table caption.
}
\seealso{
\code{\link[=SS_server]{SS_server()}}, which calls the current function
}
