% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmod_SSDayTable.R
\name{SS_DayTable_ui}
\alias{SS_DayTable_ui}
\title{User interface for the days to steady state table}
\usage{
SS_DayTable_ui(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
The user interface for the days to steady state table which includes
three outputs: a table download button, a table, and a table caption
}
\description{
User interface for the days to steady state table
}
\seealso{
\code{\link[=SS_ui]{SS_ui()}}, which calls the current function
}
