% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_outputfuncs.R
\name{SS_sol}
\alias{SS_sol}
\title{Calculate the steady state concentration solutions}
\usage{
SS_sol(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A list with steady state concentrations, steady state characteristics,
and simulation and chemical-physical parameters used in the simulation
}
\description{
This function is the main steady state module solution function from which the
server outputs are derived. This function orders the calculation of steady state
concentrations and the number of days to steady state. It formats the outputs
for appropriate display in the interface.
}
\seealso{
\code{\link[=SS_server]{SS_server()}}, which calls the current function, and \code{\link[=CalcAnalyticCss]{CalcAnalyticCss()}},
\code{\link[=CalcCssDay]{CalcCssDay()}}, and \code{\link[=StorePars_SS]{StorePars_SS()}}, which the current function calls
}
