% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmod_main.R
\name{SS_ui}
\alias{SS_ui}
\title{User interface layout for the steady state concentrations module and is composed
of three submodules}
\usage{
SS_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must match the identifier in the SS_server()}
}
\value{
User interface layout for the steady state concentrations module on the 'Run Simulation'
tab; includes one plot and two table drop down panels
}
\description{
User interface layout for the steady state concentrations module and is composed
of three submodules
}
\seealso{
\code{\link[=SS_ConcPlot_ui]{SS_ConcPlot_ui()}}, \code{\link[=SS_ConcTable_ui]{SS_ConcTable_ui()}}, and \code{\link[=SS_DayTable_ui]{SS_DayTable_ui()}},
which are all called by this function
}
