% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_outputfuncs.R
\name{StorePars_PC}
\alias{StorePars_PC}
\title{Create a data frame that stores all simulation parameters and physical-chemical
data for simulated compounds}
\usage{
StorePars_PC(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A data frame with simulation parameters and physical-chemical
data from the compounds simulated
}
\description{
This function creates a data frame that contains all simulated user-selected
parameters as well as chemical-physical compound data for simulated compounds.
Chemical-physical data is taken from the 'httk' package's
chem.physical_and_invitro.data data frame, where the data was either uploaded
by the GUI user or already present when httk was loaded.
}
