% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{textInput_DoseMatrix}
\alias{textInput_DoseMatrix}
\title{Generate a text input where users can enter a dosing regimen of times and doses
to administer of compounds selected for concentration-time profile simulations}
\usage{
textInput_DoseMatrix(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A text input box
}
\description{
Generate a text input where users can enter a dosing regimen of times and doses
to administer of compounds selected for concentration-time profile simulations
}
\examples{
textInput_DoseMatrix("multdose_odd")

}
\seealso{
\code{\link[=MS_Dosing]{MS_Dosing()}}, which calls the current function
}
