% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{textInput_OutputTimes}
\alias{textInput_OutputTimes}
\title{Generates a text input box for the user to enter concentration-time profile
output times, if desired}
\usage{
textInput_OutputTimes(id)
}
\arguments{
\item{id}{Shiny identifier name}
}
\value{
A text input box
}
\description{
Generates a text input box for the user to enter concentration-time profile
output times, if desired
}
\examples{
textInput_OutputTimes("returntimes")
}
