% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{type_func}
\alias{type_func}
\title{Determine if the data type is appropriate}
\usage{
type_func(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
TRUE or FALSE
}
\description{
Determine if the data type is appropriate
}
