% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{UpdateComps}
\alias{UpdateComps}
\title{Reset the "preloaded compounds" input selection}
\usage{
UpdateComps(session)
}
\arguments{
\item{session}{Shiny session}
}
\value{
Clear the "preloaded compounds" selection and return it to its default
value
}
\description{
This function updates the preloaded compound selections if the default to human,
in silico parameters, or model selections are changed from their original
selection.
}
\seealso{
\code{\link[=UpdateInputs]{UpdateInputs()}}, which calls the current function
}
