% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{UpdateFunc}
\alias{UpdateFunc}
\title{Reset inputs if a new main GUI function is selected}
\usage{
UpdateFunc(session)
}
\arguments{
\item{session}{Shiny session}
}
\value{
Clear "model", "in silico parameter preference", and "preloaded
compounds" selections and return to their default values
}
\description{
Reset inputs if a new main GUI function is selected
}
