% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{UpdatePars}
\alias{UpdatePars}
\title{Compile a final list of all parameters}
\usage{
UpdatePars(pars)
}
\arguments{
\item{pars}{A list of parameter values encompassing all user inputs for all modules}
}
\value{
A list of parameters to pass into modules
}
\description{
This function creates a final list of all user-selected parameters to use in
simulation. It adjusts any parameters from the user selection to a function-ready
input for all simulation modules.
}
\seealso{
\code{\link[=GatherInputVars]{GatherInputVars()}}, which calls the current function, and \code{\link[=DosingPar]{DosingPar()}},
\code{\link[=InitVals_Par]{InitVals_Par()}}, and \code{\link[=OutputTimes_Par]{OutputTimes_Par()}}, which the current function calls
}
