% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{UpdateSpec}
\alias{UpdateSpec}
\title{Reset inputs if a new species is selected}
\usage{
UpdateSpec(session)
}
\arguments{
\item{session}{Shiny session}
}
\value{
Clear "model" and "preloaded compounds" selections and return to their
default values
}
\description{
This function updates the model and preloaded compound selections if the
'species' function is changed from its original selection.
}
\seealso{
\code{\link[=UpdateInputs]{UpdateInputs()}}, which calls the current function
}
