% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{UploadComps_Check}
\alias{UploadComps_Check}
\title{Are there any formatting or data errors in the new compounds file uploaded by
the user?}
\usage{
UploadComps_Check(value, input)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}
}
\value{
If condition if FALSE, then return the message. If the condition is
TRUE, return nothing.
}
\description{
Are there any formatting or data errors in the new compounds file uploaded by
the user?
}
