% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_ValidationText.R
\name{validate_text_Common}
\alias{validate_text_Common}
\title{Check if any inputs that are common among all four modules are invalid or
missing}
\usage{
validate_text_Common(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A blank error message to the user that halts computation until the
user fixes the input error.
}
\description{
This function validates that all common user-selected parameters across all
modules are acceptable. If not, the simulator will not compute until any
errors in inputs are fixed.
}
\seealso{
\code{\link[=CompileCompLst]{CompileCompLst()}}, which calls the current function
}
