% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_ValidationText.R
\name{validate_text_IVIVE}
\alias{validate_text_IVIVE}
\title{Check if any inputs specific to the in vitro in vivo extrapolation (IVIVE)
module are invalid or missing}
\usage{
validate_text_IVIVE(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A blank error message to the user that halts computation until the
user fixes the input error.
}
\description{
This function validates that all user-selected parameters across the
IVIVE module are acceptable. If not, the simulator will not compute until any
errors in inputs are fixed.
}
\seealso{
\code{\link[=IVIVE_server]{IVIVE_server()}}, which calls the current function
}
