% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_ValidationText.R
\name{validate_text_PC}
\alias{validate_text_PC}
\title{Check if any inputs specific to the parameter calculations module are invalid
or missing}
\usage{
validate_text_PC(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A blank error message to the user that halts computation until the
user fixes the input error.
}
\description{
This function validates that all user-selected parameters across the
parameter calculations module are acceptable. If not, the simulator will not
compute until any errors in inputs are fixed.
}
\seealso{
\code{\link[=PC_server]{PC_server()}}, which calls the current function
}
