% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{AP_Bioavailability}
\alias{AP_Bioavailability}
\title{User interface for the 'Bioavailability' card on the 'Advanced Parameters' tab}
\usage{
AP_Bioavailability()
}
\value{
The 'Bioavailability' card which offers users the option to customize
five bioavailability parameters, unless the parameter calculations module
function is chosen
}
\description{
This function outputs the user interface for the 'Bioavailability' card on the
'Advanced Parameters' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=numericInput_CacoDefault]{numericInput_CacoDefault()}},
\code{\link[=selectInput_Fabs]{selectInput_Fabs()}}, \code{\link[=selectInput_Fgut]{selectInput_Fgut()}}, \code{\link[=selectInput_Overwrite]{selectInput_Overwrite()}}, and
\code{\link[=selectInput_Keep100]{selectInput_Keep100()}}, which are called by the current function
}
