% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{AP_ModelSolver}
\alias{AP_ModelSolver}
\title{User interface for the 'Model Solver' card on the 'Advanced Parameters' tab}
\usage{
AP_ModelSolver()
}
\value{
The 'Model Solver' card which allows the user options to select various
solver conditions if the concentration-time profiles module function is chosen
}
\description{
User interface for the 'Model Solver' card on the 'Advanced Parameters' tab
}
