% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{AP_OutputSpecification}
\alias{AP_OutputSpecification}
\title{User interface for the 'Output Specification' card on the 'Advanced Parameters'
tab}
\usage{
AP_OutputSpecification()
}
\value{
The 'Output Specification' card which offers the users different
selection options regarding the output for different module functions chosen
}
\description{
This function outputs the user interface for the 'Output Specification' card
on the 'Advanced Parameters' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=textInput_OutputTimes]{textInput_OutputTimes()}},
\code{\link[=selectInput_SSoutunits]{selectInput_SSoutunits()}}, \code{\link[=selectInput_OutConc]{selectInput_OutConc()}}, \code{\link[=selectInput_Tissue]{selectInput_Tissue()}},
and \code{\link[=selectInput_IVIVEoutunits]{selectInput_IVIVEoutunits()}}, which are called by the current function
}
