% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{App server function}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Input objects from the user interface}

\item{output}{Output objects from the server}

\item{session}{Session identifier}
}
\value{
Various server outputs
}
\description{
This function is the top-level server function for the ToCS application and
dictates the outputs to be displayed in the app_ui function. Calls to all other
server functions stem from this function.
}
\seealso{
\code{\link[=Model_Input]{Model_Input()}}, \code{\link[=PreloadComps_UI]{PreloadComps_UI()}}, \code{\link[=CompileCompLst]{CompileCompLst()}}, \code{\link[=UpdateInputs]{UpdateInputs()}},
\code{\link[=GatherInputVars]{GatherInputVars()}}, \code{\link[=InputRules_Children]{InputRules_Children()}}, \code{\link[=InputRules_Parents]{InputRules_Parents()}}, and
\code{\link[=Run_Simulation]{Run_Simulation()}}
}
