% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{Create_Individ_Subplot}
\alias{Create_Individ_Subplot}
\title{Title}
\usage{
Create_Individ_Subplot(i, j, comp_sol, compound_names, plt_colors, col_names)
}
\arguments{
\item{col_names}{}
}
\description{
Title
}
