% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{Create_Plotting_df}
\alias{Create_Plotting_df}
\title{Reformat the ADME solution array}
\usage{
Create_Plotting_df(i, n, sol_array, chemnames, numtimes)
}
\arguments{
\item{i}{A number representing the index of the current model state to be plotted}

\item{n}{A number representing the number of compounds being plotted}

\item{sol_array}{An array (row = time steps, column = model state, sheet = compound) of the model solution from other code}

\item{chemnames}{A vector containing the names of the compounds being simulated}

\item{numtimes}{A vector containing the times the solution is evaluated}
}
\value{
A data frame that reformats the sol_array so concentration-time profiles can be plotted according to compound
}
\description{
Reformat the ADME solution array
}
\examples{
None
}
