% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{CS_Instructions}
\alias{CS_Instructions}
\title{User interface for the 'Instructions' card on the 'Compound Selection' tab}
\usage{
CS_Instructions()
}
\value{
The 'Instructions' card with instructions for selecting and uploading
compound data, including folders with sample data
}
\description{
User interface for the 'Instructions' card on the 'Compound Selection' tab
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=Instructions_CompSelect_Part1]{Instructions_CompSelect_Part1()}},
which is called by the current function
}
