% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{CS_PreloadedCompounds}
\alias{CS_PreloadedCompounds}
\title{User interface for the 'Preloaded Compounds' card on the 'Compound Selection'
tab}
\usage{
CS_PreloadedCompounds()
}
\value{
The 'Preloaded Compounds' card that allows the user to select desired
chemicals to simulate from a list as well as any chemical-related simulation
preferences
}
\description{
User interface for the 'Preloaded Compounds' card on the 'Compound Selection'
tab
}
