% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{CS_UploadedData}
\alias{CS_UploadedData}
\title{User interface for the 'Uploaded Data' card on the 'Compound Selection' tab}
\usage{
CS_UploadedData()
}
\value{
The 'Uploaded Data' card which gives the user the option to upload a
file with new chemicals and their data. If the IVIVE module is chosen, then
users can also upload bioactivity and exposure data.
}
\description{
This function outputs the user interface for the 'UploadedData' card on the
'Compound Selection' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=fileInput_BioactiveConc]{fileInput_BioactiveConc()}}
and \code{\link[=fileInput_ExposureData]{fileInput_ExposureData()}}, which are called by the current function
}
