% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{DataTypeList}
\alias{DataTypeList}
\title{Generate a list of acceptable input data types for the compound file user
upload}
\usage{
DataTypeList()
}
\value{
A list of all acceptable data types for each column in the 'httk'
package's chem.physical_and_invitro.data
}
\description{
Generate a list of acceptable input data types for the compound file user
upload
}
