% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{ExposureUpload_Check}
\alias{ExposureUpload_Check}
\title{Title}
\usage{
ExposureUpload_Check(value, input)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}
}
\description{
Title
}
