% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{fetal_cond}
\alias{fetal_cond}
\title{Was the human species selected when running the fetal_pbtk model?}
\usage{
fetal_cond(
  value,
  input,
  message = "The 'Human' species must be selected to run the fetal_pbtk model"
)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}

\item{message}{A character string to return to the user if condition is true}
}
\value{
If condition if FALSE, then return the message. If the condition is
TRUE, return nothing.
}
\description{
Was the human species selected when running the fetal_pbtk model?
}
