% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{GatherInputVars}
\alias{GatherInputVars}
\title{Compile all parameters into one final list}
\usage{
GatherInputVars(input, compoundlist)
}
\arguments{
\item{input}{The 'input' object used as the prefix for all ID's (should be
entered as 'input')}

\item{compoundlist}{A data frame with a single column that has the alphabetized
list of compounds to simulate}
}
\value{
A named list of all parameters in the GUI to be used in the main
module functions
}
\description{
This function puts all user-selected parameters for all modules into a single
parameter list to pass through the functions that generate model outputs. The
list of parameters has each parameter in a form that the httk functions can
read them.
}
\seealso{
\code{\link[=app_server]{app_server()}}, which calls the current function, and \code{\link[=ParNames]{ParNames()}},
\code{\link[=CompilePars]{CompilePars()}}, and \code{\link[=UpdatePars]{UpdatePars()}}, which are called by the current function
}
