% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{getPiped}
\alias{getPiped}
\title{Generate the final list of available compounds to simulate in httk}
\usage{
getPiped(CASnums, honda, comptype)
}
\arguments{
\item{CASnums}{The CASRN numbers of compounds that have enough data in httk
to simulate for the user's selected parameters}

\item{honda}{The selected IVIVE assumption (either NULL, Honda1, Honda2,
Honda3, or Honda4), if applicable}
}
\value{
A vector of CASRN and compound name pairings for compounds with enough
available data to simulate in httk
}
\description{
Generate the final list of available compounds to simulate in httk
}
\examples{
getPiped(c("135410-20-7","94-82-6","30560-19-1","71751-41-2","34256-82-1"),"NULL")
getPiped(c("15972-60-8","116-06-3","834-12-8","33089-61-1","1912-24-9"),"Honda1")

}
