% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{GP_Output}
\alias{GP_Output}
\title{User interface for the 'Output' card on the 'General Parameters' tab}
\usage{
GP_Output()
}
\value{
The 'Output' card with the function selection option
}
\description{
This function outputs the user interface for the 'Output' card on the
'General Parameters' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=selectInput_Function]{selectInput_Function()}},
which is called by the current function
}
