% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{GP_Species}
\alias{GP_Species}
\title{User interface for the 'Species' card on the 'General Parameters' tab}
\usage{
GP_Species()
}
\value{
The 'Species' card with species selection options
}
\description{
This function outputs the user interface for the 'Species' card on the
'General Parameters' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=selectInput_Species]{selectInput_Species()}}
and \code{\link[=selectInput_DefaultToHuman]{selectInput_DefaultToHuman()}}, which are called by the current function
}
