% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{InitVals_Par}
\alias{InitVals_Par}
\title{Create the named vector of initial conditions needed for ADME simulations}
\usage{
InitVals_Par(model, ICopts, pars)
}
\arguments{
\item{model}{The model the user selects}

\item{ICopts}{The user's selection on whether they want to use initial
conditions other than the default, which is zero for all model states}

\item{pars}{A list of parameters to pass into modules}
}
\value{
A named vector of initial conditions for a specific model
}
\description{
Create the named vector of initial conditions needed for ADME simulations
}
\examples{
None

}
