% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{InputRules_Children}
\alias{InputRules_Children}
\title{Generate input validation rules for each individual module}
\usage{
InputRules_Children(
  iv_common,
  iv_adme,
  iv_ss,
  iv_ivive,
  iv_pc,
  input,
  ic_names
)
}
\arguments{
\item{iv_common}{A new shiny input validator object for all common inputs
across modules (function, species, default to human, model, in silico
parameters, preloaded compounds, and uploaded compounds)}

\item{iv_adme}{A new shiny input validator object for all concentration-time
profile specific parameters that other modules do not have}

\item{iv_ss}{A new shiny input validator object for all steady state
concentrations specific parameters that other modules do not have}

\item{iv_ivive}{A new shiny input validator object for all IVIVE specific
parameters that other modules do not have}

\item{iv_pc}{A new shiny input validator object for all parameter calculations
specific parameters that other modules do not have}

\item{input}{The 'input' object used as the prefix for all ID's (should be
entered as 'input')}

\item{ic_names}{A named vector of all initial condition compartment names.
See global.R and input_functions.R for more information.}
}
\value{
Validation rules attached to each child input validator
}
\description{
This function generates the rules for each input selection that the user
must follow or the program will not run. This includes not leaving parameters
blank, having them be the correct type of input, and that they're formatted
correctly. For file uploads, it checks the inputted file. Error messages are
shown to the user if the input rules are broken.
}
\seealso{
\code{\link[=app_server]{app_server()}}, which calls the current function, and \code{\link[=not_null]{not_null()}},
\code{\link[=UploadComps_Check]{UploadComps_Check()}}, \code{\link[=multdose_Select]{multdose_Select()}}, \code{\link[=multdose_odd]{multdose_odd()}}, \code{\link[=fetal_cond]{fetal_cond()}},
\code{\link[=returntimes_cond]{returntimes_cond()}}, \code{\link[=addrule_adme_ics]{addrule_adme_ics()}}, \code{\link[=BioUpload_Check]{BioUpload_Check()}}, \code{\link[=FSBf_Check]{FSBf_Check()}},
and \code{\link[=ExposureUpload_Check]{ExposureUpload_Check()}}, which are called by the current function
}
