% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_outputfuncs.R
\name{IVIVEplotting}
\alias{IVIVEplotting}
\title{Plot the oral equivalent dose (OED) and exposure data (if applicable)}
\usage{
IVIVEplotting(OED_data, BioactiveConc, pars, logscale, expdata)
}
\arguments{
\item{OED_data}{A data frame of chemical names and their OEDs (either one per
chemical or many per chemical if samples were returned); The first output
from the IVIVEsol function}

\item{BioactiveConc}{An input not used}

\item{pars}{A list of all user input parameters for the entire app}

\item{logscale}{Checkbox input value indicating if the user wanted the y-axis
of plots to be a log10 scale}

\item{expdata}{A data frame with chemical names and exposure data estimates}
}
\value{
A ggplot2 plotting object
}
\description{
Plot the oral equivalent dose (OED) and exposure data (if applicable)
}
