% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{log10breaks}
\alias{log10breaks}
\title{Generate a vector of 10^x breakpoints for the y-axis of a plot}
\usage{
log10breaks(ydata)
}
\arguments{
\item{ydata}{A vector of y-coordinate data to be plotted}
}
\value{
A vector of y-axis breaks
}
\description{
Generate a vector of 10^x breakpoints for the y-axis of a plot
}
\seealso{
\code{\link[=IVIVEplot_logscale]{IVIVEplot_logscale()}} and \code{\link[=BERplotting]{BERplotting()}}, which call the current function
}
