% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{MS_Dosing}
\alias{MS_Dosing}
\title{User interface for the 'Dosing' card on the 'Model Specifications' tab}
\usage{
MS_Dosing()
}
\value{
The 'Dosing' card with dosing selection options depending on the
module function chosen
}
\description{
User interface for the 'Dosing' card on the 'Model Specifications' tab
}
