% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{MS_Model}
\alias{MS_Model}
\title{User interface for the 'Model' card on the 'Model Specifications' tab}
\usage{
MS_Model()
}
\value{
The 'Model' card with model selection options depending on the module
function chosen
}
\description{
User interface for the 'Model' card on the 'Model Specifications' tab
}
