% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{names_ICs}
\alias{names_ICs}
\title{Generates a list of compartment names and variable names of the compartments
in the model to be used to generate initial condition options to the user
during concentration-time profile simulations}
\usage{
names_ICs()
}
\value{
A list of full compartment names and variable names of compartments
}
\description{
Generates a list of compartment names and variable names of the compartments
in the model to be used to generate initial condition options to the user
during concentration-time profile simulations
}
\examples{
names_ICs()

}
