% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{not_null}
\alias{not_null}
\title{Has the user selected and/or uploaded compounds to simulate?}
\usage{
not_null(
  value,
  input,
  message = "At least one compound must be selected or uploaded"
)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}

\item{message}{A character string to return to the user if condition is true}
}
\value{
If condition if FALSE, then return the message. If the condition is
TRUE, return nothing.
}
\description{
This function determines whether at least one compound has been set by the
user to be simulated. The compound may be either from the pre-available drop
down list or uploaded by the user.
}
\seealso{
\code{\link[=InputRules_Children]{InputRules_Children()}}, which calls the current function
}
