% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{OutputTimes_Par}
\alias{OutputTimes_Par}
\title{Generate a vector of output times}
\usage{
OutputTimes_Par(pars)
}
\arguments{
\item{pars}{A list of parameters to pass into modules}
}
\value{
A vector of times to output the ADME concentration-time profile solution
}
\description{
This function creates a vector of output times for the concentration-time profiles
module. The vector entries depend on the user's model, simulation time, and
return time specifications.
}
\seealso{
\code{\link[=UpdatePars]{UpdatePars()}}, which calls the current function
}
