% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCmod_EVPlot.R
\name{PC_EVPlot_ui}
\alias{PC_EVPlot_ui}
\title{User interface function for the parameter calculations parameter plot}
\usage{
PC_EVPlot_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in PC_EVPlot_server()}
}
\value{
User interface for the parameter plot drop down with three elements
}
\description{
This function contains the user interface elements for the parameter calculations
parameter plot. This interface itself contains three UI elements: a
download button, a plot, and a plot caption.
}
\seealso{
\code{\link[=PC_ui]{PC_ui()}}, which calls the current function
}
