% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCmod_EVTable.R
\name{PC_EVTable_server}
\alias{PC_EVTable_server}
\title{Server function for the parameter calculations parameter table}
\usage{
PC_EVTable_server(id, pc_args)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in PC_EVTable_ui()}

\item{pc_args}{A Shiny reactive list with the output of Parsol(), all shiny
parameters in pars(), and the logscale() input by the user}
}
\value{
Server outputs for the parameter table with three elements
}
\description{
Server function for the parameter calculations parameter table
}
