% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCmod_EVTable.R
\name{PC_EVTable_ui}
\alias{PC_EVTable_ui}
\title{User interface function for the parameter calculation parameter table}
\usage{
PC_EVTable_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must be the same id used as in PC_EVTable_server()}
}
\value{
User interface for the parameter table drop down with three elements
}
\description{
User interface function for the parameter calculation parameter table
}
