% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{PreloadCompsInput}
\alias{PreloadCompsInput}
\title{Generate the drop down list of preloaded compounds available to simulate in httk}
\usage{
PreloadCompsInput(
  func,
  species,
  defaulthuman,
  insilico,
  model,
  honda,
  comptype
)
}
\arguments{
\item{func}{The user-selected desired output}

\item{species}{The user-selected species}

\item{defaulthuman}{The user-selected human in vitro data preference if an
animal species is selected}

\item{insilico}{The user-selected preference for use of in silico generated
parameters if in vitro data is missing}

\item{model}{The user-selected species}

\item{honda}{The selected IVIVE assumption (either NULL, Honda1, Honda2,
Honda3, or Honda4), if applicable}
}
\value{
A drop down list of preloaded available compounds within ToCS
}
\description{
This function creates the drop down list of preloaded compounds for users to
select from under the 'Preloaded Compounds' card on the 'Compound Selection'
tab. Four sets of in silico parameters (three for clint and fup values, one
for caco2 values) will be loaded if desired by the user.
}
\examples{
PreloadCompsInput("Concentration-time profiles","Human","Yes","No, do not load in silico parameters","pbtk","NULL")
PreloadCompsInput("In vitro in vivo extrapolation (IVIVE)","Human","Yes","Yes, load in silico parameters","3compartmentss","Honda1")

}
\seealso{
\code{\link[=PreloadComps_UI]{PreloadComps_UI()}}, which calls the current function, and \code{\link[=getCASnums]{getCASnums()}}
and \code{\link[=getPiped]{getPiped()}}, wich the current function calls
}
