% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{PreloadCompsInput}
\alias{PreloadCompsInput}
\title{Generate the drop down list of preloaded compounds available to simulate in httk}
\usage{
PreloadCompsInput(func, species, defaulthuman, insilico, model, honda)
}
\arguments{
\item{func}{The user-selected desired output}

\item{species}{The user-selected species}

\item{defaulthuman}{The user-selected human in vitro data preference if an
animal species is selected}

\item{insilico}{The user-selected preference for use of in silico generated
parameters if in vitro data is missing}

\item{model}{The user-selected species}

\item{honda}{The selected IVIVE assumption (either NULL, Honda1, Honda2,
Honda3, or Honda4), if applicable}
}
\value{
A drop down list of preloaded available compounds within ToCS
}
\description{
Generate the drop down list of preloaded compounds available to simulate in httk
}
\examples{
PreloadCompsInput("Concentration-time profiles","Human","Yes","No, do not load in silico parameters","pbtk","NULL")
PreloadCompsInput("In vitro in vivo extrapolation (IVIVE)","Human","Yes","Yes, load in silico parameters","3compartmentss","Honda1")

}
