% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{Rearr_TKSumArray}
\alias{Rearr_TKSumArray}
\title{Title}
\usage{
Rearr_TKSumArray(tk_sum_array, pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\description{
Title
}
