% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{RS_Results}
\alias{RS_Results}
\title{User interface for the 'Results' card on the 'Run Simulation' tab}
\usage{
RS_Results()
}
\value{
The 'Results' card which displays a number of drop down menus filled
with output results depending on the module function chosen
}
\description{
This function outputs the user interface for the 'Results' card on the
'Run Simulation' tab of the GUI.
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, which calls the current function, and \code{\link[=ADME_ui]{ADME_ui()}},
\code{\link[=SS_ui]{SS_ui()}}, \code{\link[=IVIVE_ui]{IVIVE_ui()}}, and \code{\link[=PC_ui]{PC_ui()}}, which are called by the current function
}
