% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{run_ToCS}
\alias{run_ToCS}
\title{Use this function to run the ToCS App}
\usage{
run_ToCS()
}
\value{
A popup window with the ToCS user interface
}
\description{
This is the primary function to run the ToCS app once the ToCS package is
loaded into R. A window should open after typing 'run_ToCS' into the console.
}
\examples{
# Open the ToCS interface
run_ToCS()


}
