% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{run_ToCS}
\alias{run_ToCS}
\title{Run the ToCS App}
\usage{
run_ToCS()
}
\value{
The ToCS interface
}
\description{
This is the primary function to run the ToCS app once the ToCS package is
loaded into R or opened on the web.
}
