% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{Set_Plot_Legend_Color}
\alias{Set_Plot_Legend_Color}
\title{Remove plot legends and determine plot color}
\usage{
Set_Plot_Legend_Color(plt_lst, n_states)
}
\arguments{
\item{plt_lst}{A list of ggplot2 plotting objects where each list element is
the plot of a single compound and is composed of subplots}

\item{n_states}{Number of model compartments/outputs}
}
\value{
A list containing two elements, the first one being the inputted plot
list with the legends on each compound plot removed and a new subplot with the
plotting legend and the second element being a data frame of hexadecimal RGB
triplet colors that defines the plotting color for each compound.
}
\description{
Remove plot legends and determine plot color
}
