% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADME_outputfuncs.R
\name{SetArraySize}
\alias{SetArraySize}
\title{Set the dimensions of the arrays for the model solution and the toxicokinetic
summary data}
\usage{
SetArraySize(modelsol, n)
}
\arguments{
\item{modelsol}{The matrix resulting from solve_model, where rows represent
a time step and columns represent a model compartment}

\item{n}{The number of compounds to be simulated}
}
\value{
A list of two arrays, one for the model solution and the other
for the toxicokinetic summary data
}
\description{
This function creates the arrays (filled with zeros) that will be used to store
the outputted model solution from httk's solve_model and outputted toxicokinetic
summary information.
}
\seealso{
\code{\link[=modsol]{modsol()}}, which calls the current function
}
