% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmod_ConcTable.R
\name{SS_ConcTable_ui}
\alias{SS_ConcTable_ui}
\title{User interface for the steady state concentrations table}
\usage{
SS_ConcTable_ui(id)
}
\arguments{
\item{id}{Shiny identifier name; must match the identifier in the SS_ConcTable_server()}
}
\value{
User interface for the steady state concentrations table which includes
four elements: a table download button, a simulation parameters download button,
the steady state concentrations table, and a table caption
}
\description{
User interface for the steady state concentrations table
}
\seealso{
\code{\link[=SS_ui]{SS_ui()}}, which calls the current function
}
