% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_outputfuncs.R
\name{SS_sol}
\alias{SS_sol}
\title{Calculate the steady state concentration solutions}
\usage{
SS_sol(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A list with steady state concentrations, steady state characteristics,
and simulation and chemical-physical parameters used in the simulation
}
\description{
Calculate the steady state concentration solutions
}
