% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{UpdateComps}
\alias{UpdateComps}
\title{Reset the "preloaded compounds" input selection}
\usage{
UpdateComps(session)
}
\arguments{
\item{session}{Shiny session}
}
\value{
Clear the "preloaded compounds" selection and return it to its default
value
}
\description{
Reset the "preloaded compounds" input selection
}
