% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{UpdateInputs}
\alias{UpdateInputs}
\title{Main function to update inputs when a trigger input is changed}
\usage{
UpdateInputs(input, session)
}
\arguments{
\item{input}{The 'input' object used as the prefix for all ID's (should be
entered as 'input')}

\item{session}{Shiny session}
}
\value{
Cleared inputs for the applicable function when a main input is
changed
}
\description{
This function resets the user-selected values of certain parameters depending
on the change of another main parameters. This function ensures that the user
makes appropriate parameter selections based on the other parameters selected.
}
\seealso{
\code{\link[=app_server]{app_server()}}, which calls the current function, and \code{\link[=UpdateFunc]{UpdateFunc()}},
\code{\link[=UpdateSpec]{UpdateSpec()}}, and \code{\link[=UpdateComps]{UpdateComps()}}, which are called by the current function
}
