% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Multiple.R
\name{UpdatePars}
\alias{UpdatePars}
\title{Compile the list of all GUI parameters and adjust any parameters from the user
selection wording to a function-ready input for all simulation modules}
\usage{
UpdatePars(pars)
}
\arguments{
\item{pars}{A list of parameter values encompassing all user inputs for all modules}
}
\value{
A list of parameters to pass into modules
}
\description{
Compile the list of all GUI parameters and adjust any parameters from the user
selection wording to a function-ready input for all simulation modules
}
