% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SubFunctions.R
\name{UpdateSpec}
\alias{UpdateSpec}
\title{Reset inputs if a new species is selected}
\usage{
UpdateSpec(session)
}
\arguments{
\item{session}{Shiny session}
}
\value{
Clear "model" and "preloaded compounds" selections and return to their
default values
}
\description{
Reset inputs if a new species is selected
}
