% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorFunctions.R
\name{UploadComps_Check}
\alias{UploadComps_Check}
\title{Are there any formatting or data errors in the new compounds file uploaded by
the user?}
\usage{
UploadComps_Check(value, input)
}
\arguments{
\item{value}{A Shiny input ID corresponding to an input object}

\item{input}{The reactive inputs associated with the session}
}
\value{
If condition if FALSE, then return the message. If the condition is
TRUE, return nothing.
}
\description{
This function checks the uploaded file containing physical-chemical data for
compounds to simulate. The file is checked for correct formatting, missing
data, and correct data types.
}
\seealso{
\code{\link[=InputRules_Children]{InputRules_Children()}}, which calls the current function, and \code{\link[=MissingClintFup]{MissingClintFup()}}
and \code{\link[=DataInputType]{DataInputType()}}, which the current function calls
}
