% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_ValidationText.R
\name{validate_text_Common}
\alias{validate_text_Common}
\title{Check if any inputs that are common among all four modules are invalid or
missing}
\usage{
validate_text_Common(pars)
}
\arguments{
\item{pars}{A list of all user input parameters for the entire app}
}
\value{
A blank error message to the user that halts computation until the
user fixes the input error.
}
\description{
Check if any inputs that are common among all four modules are invalid or
missing
}
